Split vm.c into vm_http.c, vm_json.c, and vm_xml.c
vm.c was 15,217 lines due to the original amalgamated source. Extract
three self-contained subsystems into their own compilation units:

- vm_http.c (764 lines): HTTP request processing, URI parsing, cookies
- vm_json.c (806 lines): JSON encode/decode and tokenizer
- vm_xml.c (1,461 lines): XML parser engine, UTF-8 encode/decode

vm.c is now 12,119 lines. Follows the same pattern established by the
src/sx deamalgamation (4e68eac, 63b09d1).

Build system: only the Makefile OBJECTS list needed updating; the GNU
make pattern rule and NMake inference rules already cover new files in
src/ph7/. Builds and tests pass on all platforms (GNU full/tiny, MSVC).
