+++ src/ph7/vm.c
@@ -122,51 +122,7 @@ static sxi32 VmIsUnorderedCmp(ph7_value *pLeft,ph7_value *pRight)
 	}
 	return FALSE;
 }
-/*
- * Each parsed URI is recorded and stored in an instance of the following structure.
- * This structure and it's related routines are taken verbatim from the xHT project
- * [A modern embeddable HTTP engine implementing all the RFC2616 methods]
- * the xHT project is developed internally by Symisc Systems.
- */
-typedef struct SyhttpUri SyhttpUri;
-struct SyhttpUri
-{
-	SyString sHost;     /* Hostname or IP address */
-	SyString sPort;     /* Port number */
-	SyString sPath;     /* Mandatory resource path passed verbatim (Not decoded) */
-	SyString sQuery;    /* Query part */
-	SyString sFragment; /* Fragment part */
-	SyString sScheme;   /* Scheme */
-	SyString sUser;     /* Username */
-	SyString sPass;     /* Password */
-	SyString sRaw;      /* Raw URI */
-};
-/*
- * An instance of the following structure is used to record all MIME headers seen
- * during a HTTP interaction.
- * This structure and it's related routines are taken verbatim from the xHT project
- * [A modern embeddable HTTP engine implementing all the RFC2616 methods]
- * the xHT project is developed internally by Symisc Systems.
- */
-typedef struct SyhttpHeader SyhttpHeader;
-struct SyhttpHeader
-{
-	SyString sName;    /* Header name [i.e:"Content-Type","Host","User-Agent"]. NOT NUL TERMINATED */
-	SyString sValue;   /* Header values [i.e: "text/html"]. NOT NUL TERMINATED */
-};
-/*
- * Supported HTTP methods.
- */
-#define HTTP_METHOD_GET  1 /* GET */
-#define HTTP_METHOD_HEAD 2 /* HEAD */
-#define HTTP_METHOD_POST 3 /* POST */
-#define HTTP_METHOD_PUT  4 /* PUT */
-#define HTTP_METHOD_OTHR 5 /* Other HTTP methods [i.e: DELETE,TRACE,OPTIONS...]*/
-/*
- * Supported HTTP protocol version.
- */
-#define HTTP_PROTO_10 1 /* HTTP/1.0 */
-#define HTTP_PROTO_11 2 /* HTTP/1.1 */
+/* SyhttpUri, SyhttpHeader and HTTP method/protocol defines moved to ph7int.h */
 /*
  * Register a constant and it's associated expansion callback so that
  * it can be expanded from the target PHP program.
@@ -1774,7 +1730,7 @@ static ph7_value * VmExtractMemObj(
  * Extract a superglobal variable such as $_GET,$_POST,$_HEADERS,....
  * Return a pointer to the variable value on success.NULL otherwise.
  */
-static ph7_value * VmExtractSuper(
+PH7_PRIVATE ph7_value * PH7_VmExtractSuper(
 	ph7_vm *pVm,       /* Target VM */
 	const char *zName, /* Superglobal name: NOT NULL TERMINATED */
 	sxu32 nByte        /* zName length */
@@ -1799,7 +1755,7 @@ static ph7_value * VmExtractSuper(
  * Perform a raw hashmap insertion.
  * Refer to the [PH7_VmConfigure()] implementation for additional information.
  */
-static sxi32 VmHashmapInsert(
+PH7_PRIVATE sxi32 PH7_VmHashmapInsert(
 	ph7_hashmap *pMap,  /* Target hashmap  */
 	const char *zKey,   /* Entry key */
 	int nKeylen,        /* zKey length*/
@@ -1830,8 +1786,6 @@ static sxi32 VmHashmapInsert(
 	PH7_MemObjRelease(&sValue);
 	return rc;
 }
-/* Forward declaration */
-static sxi32 VmHttpProcessRequest(ph7_vm *pVm,const char *zRequest,int nByte);
 /*
  * Configure a working virtual machine instance.
  *
@@ -1998,25 +1952,25 @@ PH7_PRIVATE sxi32 PH7_VmConfigure(
 		ph7_value *pValue;
 		if( nOp == PH7_VM_CONFIG_ENV_ATTR ){
 			/* Extract the $_ENV superglobal */
-			pValue = VmExtractSuper(&(*pVm),"_ENV",sizeof("_ENV")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_ENV",sizeof("_ENV")-1);
 		}else if(nOp == PH7_VM_CONFIG_POST_ATTR ){
 			/* Extract the $_POST superglobal */
-			pValue = VmExtractSuper(&(*pVm),"_POST",sizeof("_POST")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_POST",sizeof("_POST")-1);
 		}else if(nOp == PH7_VM_CONFIG_GET_ATTR ){
 			/* Extract the $_GET superglobal */
-			pValue = VmExtractSuper(&(*pVm),"_GET",sizeof("_GET")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_GET",sizeof("_GET")-1);
 		}else if(nOp == PH7_VM_CONFIG_COOKIE_ATTR ){
 			/* Extract the $_COOKIE superglobal */
-			pValue = VmExtractSuper(&(*pVm),"_COOKIE",sizeof("_COOKIE")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_COOKIE",sizeof("_COOKIE")-1);
 		}else if(nOp == PH7_VM_CONFIG_SESSION_ATTR ){
 			/* Extract the $_SESSION superglobal */
-			pValue = VmExtractSuper(&(*pVm),"_SESSION",sizeof("_SESSION")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_SESSION",sizeof("_SESSION")-1);
 		}else if( nOp == PH7_VM_CONFIG_HEADER_ATTR ){
 			/* Extract the $_HEADER superglobale */
-			pValue = VmExtractSuper(&(*pVm),"_HEADER",sizeof("_HEADER")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_HEADER",sizeof("_HEADER")-1);
 		}else{
 			/* Extract the $_SERVER superglobal */
-			pValue = VmExtractSuper(&(*pVm),"_SERVER",sizeof("_SERVER")-1);
+			pValue = PH7_VmExtractSuper(&(*pVm),"_SERVER",sizeof("_SERVER")-1);
 		}
 		if( pValue == 0 || (pValue->iFlags & MEMOBJ_HASHMAP) == 0 ){
 			/* No such entry */
@@ -2026,7 +1980,7 @@ PH7_PRIVATE sxi32 PH7_VmConfigure(
 		/* Point to the hashmap */
 		pMap = (ph7_hashmap *)pValue->x.pOther;
 		/* Perform the insertion */
-		rc = VmHashmapInsert(pMap,zKey,-1,zValue,nLen);
+		rc = PH7_VmHashmapInsert(pMap,zKey,-1,zValue,nLen);
 		break;
 								   }
 	case PH7_VM_CONFIG_ARGV_ENTRY:{
@@ -2040,7 +1994,7 @@ PH7_PRIVATE sxi32 PH7_VmConfigure(
 			break;
 		}
 		/* Extract the $argv array */
-		pValue = VmExtractSuper(&(*pVm),"argv",sizeof("argv")-1);
+		pValue = PH7_VmExtractSuper(&(*pVm),"argv",sizeof("argv")-1);
 		if( pValue == 0 || (pValue->iFlags & MEMOBJ_HASHMAP) == 0 ){
 			/* No such entry */
 			rc = SXERR_NOTFOUND;
@@ -2050,7 +2004,7 @@ PH7_PRIVATE sxi32 PH7_VmConfigure(
 		pMap = (ph7_hashmap *)pValue->x.pOther;
 		/* Perform the insertion */
 		n = (sxu32)SyStrlen(zValue);
-		rc = VmHashmapInsert(pMap,0,0,zValue,(int)n);
+		rc = PH7_VmHashmapInsert(pMap,0,0,zValue,(int)n);
 		if( rc == SXRET_OK ){
 			if( pMap->nEntry > 1 ){
 				/* Append space separator first */
@@ -2123,7 +2077,7 @@ PH7_PRIVATE sxi32 PH7_VmConfigure(
 			nByte = (int)SyStrlen(zRequest);
 		}
 		/* Process the request */
-		rc = VmHttpProcessRequest(&(*pVm),zRequest,nByte);
+		rc = PH7_VmHttpProcessRequest(&(*pVm),zRequest,nByte);
 		break;
 									}
 	default:
@@ -10187,8 +10141,6 @@ static int vm_builtin_ph7_credits(ph7_context *pCtx,int nArg,ph7_value **apArg)
  * Status:
  *    Stable.
  */
-/* Forward declaration */
-static sxi32 VmHttpSplitURI(SyhttpUri *pOut,const char *zUri,sxu32 nLen);
 /*
  * value parse_url(string $url [, int $component = -1 ])
  *  Parse a URL and return its fields.
@@ -10237,7 +10189,7 @@ static int vm_builtin_parse_url(ph7_context *pCtx,int nArg,ph7_value **apArg)
 		return PH7_OK;
 	}
 	/* Get a parse */
-	rc = VmHttpSplitURI(&sURI,zStr,(sxu32)nLen);
+	rc = PH7_VmHttpSplitURI(&sURI,zStr,(sxu32)nLen);
 	if( rc != SXRET_OK ){
 		/* Malformed input,return FALSE */
 		ph7_result_bool(pCtx,0);
@@ -10758,13 +10710,13 @@ static int vm_builtin_import_request_variables(ph7_context *pCtx,int nArg,ph7_va
 		pSuper = 0;
 		if( c == 'G' || c == 'g' ){
 			/* Import $_GET variables */
-			pSuper = VmExtractSuper(pVm,"_GET",sizeof("_GET")-1);
+			pSuper = PH7_VmExtractSuper(pVm,"_GET",sizeof("_GET")-1);
 		}else if( c == 'P' || c == 'p' ){
 			/* Import $_POST variables */
-			pSuper = VmExtractSuper(pVm,"_POST",sizeof("_POST")-1);
+			pSuper = PH7_VmExtractSuper(pVm,"_POST",sizeof("_POST")-1);
 		}else if( c == 'c' || c == 'C' ){
 			/* Import $_COOKIE variables */
-			pSuper = VmExtractSuper(pVm,"_COOKIE",sizeof("_COOKIE")-1);
+			pSuper = PH7_VmExtractSuper(pVm,"_COOKIE",sizeof("_COOKIE")-1);
 		}
 		if( pSuper ){
 			/* Iterate throw array entries */
@@ -11608,2301 +11560,8 @@ static int VmProcessLongOpt(ph7_value *pKey,ph7_value *pValue,void *pUserData)
 	VmExtractOptArgValue(pOpt->pArray,pOpt->pWorker,zArg,pOpt->zArgEnd,need_value,pOpt->pCtx,zOpt);
 	return PH7_OK;
 }
-/*
- * Section:
- *  JSON encoding/decoding routines.
- * Status:
- *    Devel.
- */
-/* Forward reference */
-static int VmJsonArrayEncode(ph7_value *pKey,ph7_value *pValue,void *pUserData);
-static int VmJsonObjectEncode(const char *zAttr,ph7_value *pValue,void *pUserData);
-/*
- * JSON encoder state is stored in an instance
- * of the following structure.
- */
-typedef struct json_private_data json_private_data;
-struct json_private_data
-{
-	ph7_context *pCtx; /* Call context */
-	int isFirst;       /* True if first encoded entry */
-	int iFlags;        /* JSON encoding flags */
-	int nRecCount;     /* Recursion count */
-};
-/*
- * Returns the JSON representation of a value.In other word perform a JSON encoding operation.
- * According to wikipedia
- * JSON's basic types are:
- *   Number (double precision floating-point format in JavaScript, generally depends on implementation)
- *   String (double-quoted Unicode, with backslash escaping)
- *   Boolean (true or false)
- *   Array (an ordered sequence of values, comma-separated and enclosed in square brackets; the values
- *    do not need to be of the same type)
- *   Object (an unordered collection of key:value pairs with the ':' character separating the key
- *     and the value, comma-separated and enclosed in curly braces; the keys must be strings and should
- *     be distinct from each other)
- *   null (empty)
- * Non-significant white space may be added freely around the "structural characters"
- * (i.e. the brackets "[{]}", colon ":" and comma ",").
- */
-static sxi32 VmJsonEncode(
-	ph7_value *pIn,          /* Encode this value */
-	json_private_data *pData /* Context data */
-	){
-		ph7_context *pCtx = pData->pCtx;
-		int iFlags = pData->iFlags;
-		int nByte;
-		if( ph7_value_is_null(pIn) || ph7_value_is_resource(pIn)){
-			/* null */
-			ph7_result_string(pCtx,"null",(int)sizeof("null")-1);
-		}else if( ph7_value_is_bool(pIn) ){
-			int iBool = ph7_value_to_bool(pIn);
-			int iLen;
-			/* true/false */
-			iLen = iBool ? (int)sizeof("true") : (int)sizeof("false");
-			ph7_result_string(pCtx,iBool ? "true" : "false",iLen-1);
-		}else if(  ph7_value_is_numeric(pIn) && !ph7_value_is_string(pIn) ){
-			const char *zNum;
-			/* Get a string representation of the number */
-			zNum = ph7_value_to_string(pIn,&nByte);
-			ph7_result_string(pCtx,zNum,nByte);
-		}else if( ph7_value_is_string(pIn) ){
-			if( (iFlags & JSON_NUMERIC_CHECK) &&  ph7_value_is_numeric(pIn) ){
-				const char *zNum;
-				/* Encodes numeric strings as numbers. */
-				PH7_MemObjToReal(pIn); /* Force a numeric cast */
-				/* Get a string representation of the number */
-				zNum = ph7_value_to_string(pIn,&nByte);
-				ph7_result_string(pCtx,zNum,nByte);
-			}else{
-				const char *zIn,*zEnd;
-				int c;
-				/* Encode the string */
-				zIn = ph7_value_to_string(pIn,&nByte);
-				zEnd = &zIn[nByte];
-				/* Append the double quote */
-				ph7_result_string(pCtx,"\"",(int)sizeof(char));
-				for(;;){
-					if( zIn >= zEnd ){
-						/* No more input to process */
-						break;
-					}
-					c = zIn[0];
-					/* Advance the stream cursor */
-					zIn++;
-					if( (c == '<' || c == '>') && (iFlags & JSON_HEX_TAG) ){
-						/* All < and > are converted to \u003C and \u003E */
-						if( c == '<' ){
-							ph7_result_string(pCtx,"\\u003C",(int)sizeof("\\u003C")-1);
-						}else{
-							ph7_result_string(pCtx,"\\u003E",(int)sizeof("\\u003E")-1);
-						}
-						continue;
-					}else if( c == '&' && (iFlags & JSON_HEX_AMP) ){
-						/* All &s are converted to \u0026.  */
-						ph7_result_string(pCtx,"\\u0026",(int)sizeof("\\u0026")-1);
-						continue;
-					}else if( c == '\'' && (iFlags & JSON_HEX_APOS) ){
-						/* All ' are converted to \u0027.   */
-						ph7_result_string(pCtx,"\\u0027",(int)sizeof("\\u0027")-1);
-						continue;
-					}else if( c == '"' && (iFlags & JSON_HEX_QUOT) ){
-						/* All " are converted to \u0022. */
-						ph7_result_string(pCtx,"\\u0022",(int)sizeof("\\u0022")-1);
-						continue;
-					}
-					if( c == '"' || (c == '\\' && ((iFlags & JSON_UNESCAPED_SLASHES)==0)) ){
-						/* Unescape the character */
-						ph7_result_string(pCtx,"\\",(int)sizeof(char));
-					}
-					/* Append character verbatim */
-					ph7_result_string(pCtx,(const char *)&c,(int)sizeof(char));
-				}
-				/* Append the double quote */
-				ph7_result_string(pCtx,"\"",(int)sizeof(char));
-			}
-		}else if( ph7_value_is_array(pIn) ){
-			int c = '[',d = ']';
-			/* Encode the array */
-			pData->isFirst = 1;
-			if( iFlags & JSON_FORCE_OBJECT ){
-				/* Outputs an object rather than an array */
-				c = '{';
-				d = '}';
-			}
-			/* Append the square bracket or curly braces */
-			ph7_result_string(pCtx,(const char *)&c,(int)sizeof(char));
-			/* Iterate throw array entries */
-			ph7_array_walk(pIn,VmJsonArrayEncode,pData);
-			/* Append the closing square bracket or curly braces */
-			ph7_result_string(pCtx,(const char *)&d,(int)sizeof(char));
-		}else if( ph7_value_is_object(pIn) ){
-			/* Encode the class instance */
-			pData->isFirst = 1;
-			/* Append the curly braces */
-			ph7_result_string(pCtx,"{",(int)sizeof(char));
-			/* Iterate throw class attribute */
-			ph7_object_walk(pIn,VmJsonObjectEncode,pData);
-			/* Append the closing curly braces  */
-			ph7_result_string(pCtx,"}",(int)sizeof(char));
-		}else{
-			/* Can't happen */
-			ph7_result_string(pCtx,"null",(int)sizeof("null")-1);
-		}
-		/* All done */
-		return PH7_OK;
-}
-/*
- * The following walker callback is invoked each time we need
- * to encode an array to JSON.
- */
-static int VmJsonArrayEncode(ph7_value *pKey,ph7_value *pValue,void *pUserData)
-{
-	json_private_data *pJson = (json_private_data *)pUserData;
-	if( pJson->nRecCount > 31 ){
-		/* Recursion limit reached,return immediately */
-		return PH7_OK;
-	}
-	if( !pJson->isFirst ){
-		/* Append the colon first */
-		ph7_result_string(pJson->pCtx,",",(int)sizeof(char));
-	}
-	if( pJson->iFlags & JSON_FORCE_OBJECT ){
-		/* Outputs an object rather than an array */
-		const char *zKey;
-		int nByte;
-		/* Extract a string representation of the key */
-		zKey = ph7_value_to_string(pKey,&nByte);
-		/* Append the key and the double colon */
-		ph7_result_string_format(pJson->pCtx,"\"%.*s\":",nByte,zKey);
-	}
-	/* Encode the value */
-	pJson->nRecCount++;
-	VmJsonEncode(pValue,pJson);
-	pJson->nRecCount--;
-	pJson->isFirst = 0;
-	return PH7_OK;
-}
-/*
- * The following walker callback is invoked each time we need to encode
- * a class instance [i.e: Object in the PHP jargon] to JSON.
- */
-static int VmJsonObjectEncode(const char *zAttr,ph7_value *pValue,void *pUserData)
-{
-	json_private_data *pJson = (json_private_data *)pUserData;
-	if( pJson->nRecCount > 31 ){
-		/* Recursion limit reached,return immediately */
-		return PH7_OK;
-	}
-	if( !pJson->isFirst ){
-		/* Append the colon first */
-		ph7_result_string(pJson->pCtx,",",(int)sizeof(char));
-	}
-	/* Append the attribute name and the double colon first */
-	ph7_result_string_format(pJson->pCtx,"\"%s\":",zAttr);
-	/* Encode the value */
-	pJson->nRecCount++;
-	VmJsonEncode(pValue,pJson);
-	pJson->nRecCount--;
-	pJson->isFirst = 0;
-	return PH7_OK;
-}
-/*
- * string json_encode(mixed $value [, int $options = 0 ])
- *  Returns a string containing the JSON representation of value.
- * Parameters
- *  $value
- *  The value being encoded. Can be any type except a resource.
- * $options
- *  Bitmask consisting of:
- *  JSON_HEX_TAG   All < and > are converted to \u003C and \u003E.
- *  JSON_HEX_AMP   All &s are converted to \u0026.
- *  JSON_HEX_APOS  All ' are converted to \u0027.
- *  JSON_HEX_QUOT  All " are converted to \u0022.
- *  JSON_FORCE_OBJECT  Outputs an object rather than an array.
- *  JSON_NUMERIC_CHECK Encodes numeric strings as numbers.
- *  JSON_BIGINT_AS_STRING   Not used
- *  JSON_PRETTY_PRINT       Use whitespace in returned data to format it.
- *  JSON_UNESCAPED_SLASHES  Don't escape '/'
- *  JSON_UNESCAPED_UNICODE  Not used.
- * Return
- *  Returns a JSON encoded string on success. FALSE otherwise
- */
-static int vm_builtin_json_encode(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	json_private_data sJson;
-	if( nArg < 1 ){
-		/* Missing arguments,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Prepare the JSON data */
-	sJson.nRecCount = 0;
-	sJson.pCtx = pCtx;
-	sJson.isFirst = 1;
-	sJson.iFlags = 0;
-	if( nArg > 1 && ph7_value_is_int(apArg[1]) ){
-		/* Extract option flags */
-		sJson.iFlags = ph7_value_to_int(apArg[1]);
-	}
-	/* Perform the encoding operation */
-	VmJsonEncode(apArg[0],&sJson);
-	/* All done */
-	return PH7_OK;
-}
-/*
- * int json_last_error(void)
- *  Returns the last error (if any) occurred during the last JSON encoding/decoding.
- * Parameters
- *  None
- * Return
- *  Returns an integer, the value can be one of the following constants:
- *  JSON_ERROR_NONE            No error has occurred.
- *  JSON_ERROR_DEPTH           The maximum stack depth has been exceeded.
- *  JSON_ERROR_STATE_MISMATCH  Invalid or malformed JSON.
- *  JSON_ERROR_CTRL_CHAR  	   Control character error, possibly incorrectly encoded.
- *  JSON_ERROR_SYNTAX          Syntax error.
- *  JSON_ERROR_UTF8_CHECK      Malformed UTF-8 characters.
- */
-static int vm_builtin_json_last_error(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	/* Return the error code */
-	ph7_result_int(pCtx,pVm->json_rc);
-	SXUNUSED(nArg); /* cc warning */
-	SXUNUSED(apArg);
-	return PH7_OK;
-}
-/* Possible tokens from the JSON tokenization process */
-#define JSON_TK_TRUE    0x001 /* Boolean true */
-#define JSON_TK_FALSE   0x002 /* Boolean false */
-#define JSON_TK_STR     0x004 /* String enclosed in double quotes */
-#define JSON_TK_NULL    0x008 /* null */
-#define JSON_TK_NUM     0x010 /* Numeric */
-#define JSON_TK_OCB     0x020 /* Open curly braces '{' */
-#define JSON_TK_CCB     0x040 /* Closing curly braces '}' */
-#define JSON_TK_OSB     0x080 /* Open square bracke '[' */
-#define JSON_TK_CSB     0x100 /* Closing square bracket ']' */
-#define JSON_TK_COLON   0x200 /* Single colon ':' */
-#define JSON_TK_COMMA   0x400 /* Single comma ',' */
-#define JSON_TK_INVALID 0x800 /* Unexpected token */
-/*
- * Tokenize an entire JSON input.
- * Get a single low-level token from the input file.
- * Update the stream pointer so that it points to the first
- * character beyond the extracted token.
- */
-static sxi32 VmJsonTokenize(SyStream *pStream,SyToken *pToken,void *pUserData,void *pCtxData)
-{
-	int *pJsonErr = (int *)pUserData;
-	SyString *pStr;
-	int c;
-	/* Ignore leading white spaces */
-	while( pStream->zText < pStream->zEnd && pStream->zText[0] < 0xc0 && SyisSpace(pStream->zText[0]) ){
-		/* Advance the stream cursor */
-		if( pStream->zText[0] == '\n' ){
-			/* Update line counter */
-			pStream->nLine++;
-		}
-		pStream->zText++;
-	}
-	if( pStream->zText >= pStream->zEnd ){
-		/* End of input reached */
-		SXUNUSED(pCtxData); /* cc warning */
-		return SXERR_EOF;
-	}
-	/* Record token starting position and line */
-	pToken->nLine = pStream->nLine;
-	pToken->pUserData = 0;
-	pStr = &pToken->sData;
-	SyStringInitFromBuf(pStr,pStream->zText,0);
-	if( pStream->zText[0] == '{' || pStream->zText[0] == '[' || pStream->zText[0] == '}' || pStream->zText[0] == ']'
-		|| pStream->zText[0] == ':' || pStream->zText[0] == ',' ){
-			/* Single character */
-			c = pStream->zText[0];
-			/* Set token type */
-			switch(c){
-			case '[': pToken->nType = JSON_TK_OSB;   break;
-			case '{': pToken->nType = JSON_TK_OCB;   break;
-			case '}': pToken->nType = JSON_TK_CCB;   break;
-			case ']': pToken->nType = JSON_TK_CSB;   break;
-			case ':': pToken->nType = JSON_TK_COLON; break;
-			case ',': pToken->nType = JSON_TK_COMMA; break;
-			default:
-				break;
-			}
-			/* Advance the stream cursor */
-			pStream->zText++;
-	}else if( pStream->zText[0] == '"') {
-		/* JSON string */
-		pStream->zText++;
-		pStr->zString++;
-		/* Delimit the string */
-		while( pStream->zText < pStream->zEnd ){
-			if( pStream->zText[0] == '"' && pStream->zText[-1] != '\\' ){
-				break;
-			}
-			if( pStream->zText[0] == '\n' ){
-				/* Update line counter */
-				pStream->nLine++;
-			}
-			pStream->zText++;
-		}
-		if( pStream->zText >= pStream->zEnd ){
-			/* Missing closing '"' */
-			pToken->nType = JSON_TK_INVALID;
-			*pJsonErr = JSON_ERROR_SYNTAX;
-		}else{
-			pToken->nType = JSON_TK_STR;
-			pStream->zText++; /* Jump the closing double quotes */
-		}
-	}else if( pStream->zText[0] < 0xc0 && SyisDigit(pStream->zText[0]) ){
-		/* Number */
-		pStream->zText++;
-		pToken->nType = JSON_TK_NUM;
-		while( pStream->zText < pStream->zEnd && pStream->zText[0] < 0xc0 && SyisDigit(pStream->zText[0]) ){
-			pStream->zText++;
-		}
-		if( pStream->zText < pStream->zEnd ){
-			c = pStream->zText[0];
-			if( c == '.' ){
-					/* Real number */
-					pStream->zText++;
-					while( pStream->zText < pStream->zEnd && pStream->zText[0] < 0xc0 && SyisDigit(pStream->zText[0]) ){
-						pStream->zText++;
-					}
-					if( pStream->zText < pStream->zEnd ){
-						c = pStream->zText[0];
-						if( c=='e' || c=='E' ){
-							pStream->zText++;
-							if( pStream->zText < pStream->zEnd ){
-								c = pStream->zText[0];
-								if( c =='+' || c=='-' ){
-									pStream->zText++;
-								}
-								while( pStream->zText < pStream->zEnd && pStream->zText[0] < 0xc0 && SyisDigit(pStream->zText[0]) ){
-									pStream->zText++;
-								}
-							}
-						}
-					}
-				}else if( c=='e' || c=='E' ){
-					/* Real number */
-					pStream->zText++;
-					if( pStream->zText < pStream->zEnd ){
-						c = pStream->zText[0];
-						if( c =='+' || c=='-' ){
-							pStream->zText++;
-						}
-						while( pStream->zText < pStream->zEnd && pStream->zText[0] < 0xc0 && SyisDigit(pStream->zText[0]) ){
-							pStream->zText++;
-						}
-					}
-				}
-			}
-	}else if( XLEX_IN_LEN(pStream) >= sizeof("true") -1 &&
-		SyStrnicmp((const char *)pStream->zText,"true",sizeof("true")-1) == 0 ){
-			/* boolean true */
-			pToken->nType = JSON_TK_TRUE;
-			/* Advance the stream cursor */
-			pStream->zText += sizeof("true")-1;
-	}else if( XLEX_IN_LEN(pStream) >= sizeof("false") -1 &&
-		SyStrnicmp((const char *)pStream->zText,"false",sizeof("false")-1) == 0 ){
-			/* boolean false */
-			pToken->nType = JSON_TK_FALSE;
-			/* Advance the stream cursor */
-			pStream->zText += sizeof("false")-1;
-	}else if( XLEX_IN_LEN(pStream) >= sizeof("null") -1 &&
-		SyStrnicmp((const char *)pStream->zText,"null",sizeof("null")-1) == 0 ){
-			/* NULL */
-			pToken->nType = JSON_TK_NULL;
-			/* Advance the stream cursor */
-			pStream->zText += sizeof("null")-1;
-	}else{
-		/* Unexpected token */
-		pToken->nType = JSON_TK_INVALID;
-		/* Advance the stream cursor */
-		pStream->zText++;
-		*pJsonErr = JSON_ERROR_SYNTAX;
-		/* Abort processing immediatley */
-		return SXERR_ABORT;
-	}
-	/* record token length */
-	pStr->nByte = (sxu32)((const char *)pStream->zText-pStr->zString);
-	if( pToken->nType == JSON_TK_STR ){
-		pStr->nByte--;
-	}
-	/* Return to the lexer */
-	return SXRET_OK;
-}
-/*
- * JSON decoded input consumer callback signature.
- */
-typedef int (*ProcJsonConsumer)(ph7_context *,ph7_value *,ph7_value *,void *);
-/*
- * JSON decoder state is kept in the following structure.
- */
-typedef struct json_decoder json_decoder;
-struct json_decoder
-{
-	ph7_context *pCtx; /* Call context */
-	ProcJsonConsumer xConsumer; /* Consumer callback */
-	void *pUserData;   /* Last argument to xConsumer() */
-	int iFlags;        /* Configuration flags */
-	SyToken *pIn;      /* Token stream */
-	SyToken *pEnd;     /* End of the token stream */
-	int rec_depth;     /* Recursion limit */
-	int rec_count;     /* Current nesting level */
-	int *pErr;         /* JSON decoding error if any */
-};
-#define JSON_DECODE_ASSOC 0x01 /* Decode a JSON object as an associative array */
-/* Forward declaration */
-static int VmJsonArrayDecoder(ph7_context *pCtx,ph7_value *pKey,ph7_value *pWorker,void *pUserData);
-/*
- * Dequote [i.e: Resolve all backslash escapes ] a JSON string and store
- * the result in the given ph7_value.
- */
-static void VmJsonDequoteString(const SyString *pStr,ph7_value *pWorker)
-{
-	const char *zIn = pStr->zString;
-	const char *zEnd = &pStr->zString[pStr->nByte];
-	const char *zCur;
-	int c;
-	/* Mark the value as a string */
-	ph7_value_string(pWorker,"",0); /* Empty string */
-	for(;;){
-		zCur = zIn;
-		while( zIn < zEnd && zIn[0] != '\\' ){
-			zIn++;
-		}
-		if( zIn > zCur ){
-			/* Append chunk verbatim */
-			ph7_value_string(pWorker,zCur,(int)(zIn-zCur));
-		}
-		zIn++;
-		if( zIn >= zEnd ){
-			/* End of the input reached */
-			break;
-		}
-		c = zIn[0];
-		/* Unescape the character */
-		switch(c){
-		case '"':  ph7_value_string(pWorker,(const char *)&c,(int)sizeof(char)); break;
-		case '\\': ph7_value_string(pWorker,(const char *)&c,(int)sizeof(char)); break;
-		case 'n':  ph7_value_string(pWorker,"\n",(int)sizeof(char)); break;
-		case 'r':  ph7_value_string(pWorker,"\r",(int)sizeof(char)); break;
-		case 't':  ph7_value_string(pWorker,"\t",(int)sizeof(char)); break;
-		case 'f':  ph7_value_string(pWorker,"\f",(int)sizeof(char)); break;
-		default:
-			ph7_value_string(pWorker,(const char *)&c,(int)sizeof(char));
-			break;
-		}
-		/* Advance the stream cursor */
-		zIn++;
-	}
-}
-/*
- * Returns a ph7_value holding the image of a JSON string. In other word perform a JSON decoding operation.
- * According to wikipedia
- * JSON's basic types are:
- *   Number (double precision floating-point format in JavaScript, generally depends on implementation)
- *   String (double-quoted Unicode, with backslash escaping)
- *   Boolean (true or false)
- *   Array (an ordered sequence of values, comma-separated and enclosed in square brackets; the values
- *    do not need to be of the same type)
- *   Object (an unordered collection of key:value pairs with the ':' character separating the key
- *     and the value, comma-separated and enclosed in curly braces; the keys must be strings and should
- *     be distinct from each other)
- *   null (empty)
- * Non-significant white space may be added freely around the "structural characters" (i.e. the brackets "[{]}", colon ":" and comma ",").
- */
-static sxi32 VmJsonDecode(
-	json_decoder *pDecoder, /* JSON decoder */
-	ph7_value *pArrayKey    /* Key for the decoded array */
-	){
-	ph7_value *pWorker; /* Worker variable */
-	sxi32 rc;
-	/* Check if we do not nest to much */
-	if( pDecoder->rec_count >= pDecoder->rec_depth ){
-		/* Nesting limit reached,abort decoding immediately */
-		*pDecoder->pErr = JSON_ERROR_DEPTH;
-		return SXERR_ABORT;
-	}
-	if( pDecoder->pIn->nType & (JSON_TK_STR|JSON_TK_TRUE|JSON_TK_FALSE|JSON_TK_NULL|JSON_TK_NUM) ){
-		/* Scalar value */
-		pWorker = ph7_context_new_scalar(pDecoder->pCtx);
-		if( pWorker == 0 ){
-			ph7_context_throw_error(pDecoder->pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-			/* Abort the decoding operation immediately */
-			return SXERR_ABORT;
-		}
-		/* Reflect the JSON image */
-		if( pDecoder->pIn->nType & JSON_TK_NULL ){
-			/* Nullify the value.*/
-			ph7_value_null(pWorker);
-		}else if( pDecoder->pIn->nType & (JSON_TK_TRUE|JSON_TK_FALSE) ){
-			/* Boolean value */
-			ph7_value_bool(pWorker,(pDecoder->pIn->nType & JSON_TK_TRUE) ? 1 : 0 );
-		}else if( pDecoder->pIn->nType & JSON_TK_NUM ){
-			SyString *pStr = &pDecoder->pIn->sData;
-			/*
-			 * Numeric value.
-			 * Get a string representation first then try to get a numeric
-			 * value.
-			 */
-			ph7_value_string(pWorker,pStr->zString,(int)pStr->nByte);
-			/* Obtain a numeric representation */
-			PH7_MemObjToNumeric(pWorker);
-		}else{
-			/* Dequote the string */
-			VmJsonDequoteString(&pDecoder->pIn->sData,pWorker);
-		}
-		/* Invoke the consumer callback */
-		rc = pDecoder->xConsumer(pDecoder->pCtx,pArrayKey,pWorker,pDecoder->pUserData);
-		if( rc == SXERR_ABORT ){
-			return SXERR_ABORT;
-		}
-		/* All done,advance the stream cursor */
-		pDecoder->pIn++;
-	}else if( pDecoder->pIn->nType & JSON_TK_OSB /*'[' */) {
-		ProcJsonConsumer xOld;
-		void *pOld;
-		/* Array representation*/
-		pDecoder->pIn++;
-		/* Create a working array */
-		pWorker = ph7_context_new_array(pDecoder->pCtx);
-		if( pWorker == 0 ){
-			ph7_context_throw_error(pDecoder->pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-			/* Abort the decoding operation immediately */
-			return SXERR_ABORT;
-		}
-		/* Save the old consumer */
-		xOld = pDecoder->xConsumer;
-		pOld = pDecoder->pUserData;
-		/* Set the new consumer */
-		pDecoder->xConsumer = VmJsonArrayDecoder;
-		pDecoder->pUserData = pWorker;
-		/* Decode the array */
-		for(;;){
-			/* Jump trailing comma. Note that the standard PHP engine will not let you
-			 * do this.
-			 */
-			while( (pDecoder->pIn < pDecoder->pEnd) && (pDecoder->pIn->nType & JSON_TK_COMMA) ){
-				pDecoder->pIn++;
-			}
-			if( pDecoder->pIn >= pDecoder->pEnd || (pDecoder->pIn->nType & JSON_TK_CSB) /*']'*/ ){
-				if( pDecoder->pIn < pDecoder->pEnd ){
-					pDecoder->pIn++; /* Jump the trailing ']' */
-				}
-				break;
-			}
-			/* Recurse and decode the entry */
-			pDecoder->rec_count++;
-			rc = VmJsonDecode(pDecoder,0);
-			pDecoder->rec_count--;
-			if( rc == SXERR_ABORT ){
-				/* Abort processing immediately */
-				return SXERR_ABORT;
-			}
-			/*The cursor is automatically advanced by the VmJsonDecode() function */
-			if( (pDecoder->pIn < pDecoder->pEnd) &&
-				((pDecoder->pIn->nType & (JSON_TK_CSB/*']'*/|JSON_TK_COMMA/*','*/))==0) ){
-					/* Unexpected token,abort immediatley */
-					*pDecoder->pErr = JSON_ERROR_SYNTAX;
-					return SXERR_ABORT;
-			}
-		}
-		/* Restore the old consumer */
-		pDecoder->xConsumer = xOld;
-		pDecoder->pUserData = pOld;
-		/* Invoke the old consumer on the decoded array */
-		xOld(pDecoder->pCtx,pArrayKey,pWorker,pOld);
-	}else if( pDecoder->pIn->nType & JSON_TK_OCB /*'{' */) {
-		ProcJsonConsumer xOld;
-		ph7_value *pKey;
-		void *pOld;
-		/* Object representation*/
-		pDecoder->pIn++;
-		/* Return the object as an associative array */
-		if( (pDecoder->iFlags & JSON_DECODE_ASSOC) == 0 ){
-			ph7_context_throw_error(pDecoder->pCtx,PH7_CTX_WARNING,
-				"JSON Objects are always returned as an associative array"
-				);
-		}
-		/* Create a working array */
-		pWorker = ph7_context_new_array(pDecoder->pCtx);
-		pKey = ph7_context_new_scalar(pDecoder->pCtx);
-		if( pWorker == 0 || pKey == 0){
-			ph7_context_throw_error(pDecoder->pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-			/* Abort the decoding operation immediately */
-			return SXERR_ABORT;
-		}
-		/* Save the old consumer */
-		xOld = pDecoder->xConsumer;
-		pOld = pDecoder->pUserData;
-		/* Set the new consumer */
-		pDecoder->xConsumer = VmJsonArrayDecoder;
-		pDecoder->pUserData = pWorker;
-		/* Decode the object */
-		for(;;){
-			/* Jump trailing comma. Note that the standard PHP engine will not let you
-			 * do this.
-			 */
-			while( (pDecoder->pIn < pDecoder->pEnd) && (pDecoder->pIn->nType & JSON_TK_COMMA) ){
-				pDecoder->pIn++;
-			}
-			if( pDecoder->pIn >= pDecoder->pEnd || (pDecoder->pIn->nType & JSON_TK_CCB) /*'}'*/ ){
-				if( pDecoder->pIn < pDecoder->pEnd ){
-					pDecoder->pIn++; /* Jump the trailing ']' */
-				}
-				break;
-			}
-			if( (pDecoder->pIn->nType & JSON_TK_STR) == 0 || &pDecoder->pIn[1] >= pDecoder->pEnd
-				|| (pDecoder->pIn[1].nType & JSON_TK_COLON) == 0){
-					/* Syntax error,return immediately */
-					*pDecoder->pErr = JSON_ERROR_SYNTAX;
-					return SXERR_ABORT;
-			}
-			/* Dequote the key */
-			VmJsonDequoteString(&pDecoder->pIn->sData,pKey);
-			/* Jump the key and the colon */
-			pDecoder->pIn += 2;
-			/* Recurse and decode the value */
-			pDecoder->rec_count++;
-			rc = VmJsonDecode(pDecoder,pKey);
-			pDecoder->rec_count--;
-			if( rc == SXERR_ABORT ){
-				/* Abort processing immediately */
-				return SXERR_ABORT;
-			}
-			/* Reset the internal buffer of the key */
-			ph7_value_reset_string_cursor(pKey);
-			/*The cursor is automatically advanced by the VmJsonDecode() function */
-		}
-		/* Restore the old consumer */
-		pDecoder->xConsumer = xOld;
-		pDecoder->pUserData = pOld;
-		/* Invoke the old consumer on the decoded object*/
-		xOld(pDecoder->pCtx,pArrayKey,pWorker,pOld);
-		/* Release the key */
-		ph7_context_release_value(pDecoder->pCtx,pKey);
-	}else{
-		/* Unexpected token */
-		return SXERR_ABORT; /* Abort immediately */
-	}
-	/* Release the worker variable */
-	ph7_context_release_value(pDecoder->pCtx,pWorker);
-	return SXRET_OK;
-}
-/*
- * The following JSON decoder callback is invoked each time
- * a JSON array representation [i.e: [15,"hello",FALSE] ]
- * is being decoded.
- */
-static int VmJsonArrayDecoder(ph7_context *pCtx,ph7_value *pKey,ph7_value *pWorker,void *pUserData)
-{
-	ph7_value *pArray = (ph7_value *)pUserData;
-	/* Insert the entry */
-	ph7_array_add_elem(pArray,pKey,pWorker); /* Will make it's own copy */
-	SXUNUSED(pCtx); /* cc warning */
-	/* All done */
-	return SXRET_OK;
-}
-/*
- * Standard JSON decoder callback.
- */
-static int VmJsonDefaultDecoder(ph7_context *pCtx,ph7_value *pKey,ph7_value *pWorker,void *pUserData)
-{
-	/* Return the value directly */
-	ph7_result_value(pCtx,pWorker); /* Will make it's own copy */
-	SXUNUSED(pKey); /* cc warning */
-	SXUNUSED(pUserData);
-	/* All done */
-	return SXRET_OK;
-}
-/*
- * mixed json_decode(string $json[,bool $assoc = false[,int $depth = 32[,int $options = 0 ]]])
- *  Takes a JSON encoded string and converts it into a PHP variable.
- * Parameters
- *  $json
- *    The json string being decoded.
- * $assoc
- *   When TRUE, returned objects will be converted into associative arrays.
- * $depth
- *   User specified recursion depth.
- * $options
- *   Bitmask of JSON decode options. Currently only JSON_BIGINT_AS_STRING is supported
- * (default is to cast large integers as floats)
- * Return
- *  The value encoded in json in appropriate PHP type. Values true, false and null (case-insensitive)
- *  are returned as TRUE, FALSE and NULL respectively. NULL is returned if the json cannot be decoded
- *  or if the encoded data is deeper than the recursion limit.
- */
-static int vm_builtin_json_decode(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_vm *pVm = pCtx->pVm;
-	json_decoder sDecoder;
-	const char *zIn;
-	SySet sToken;
-	SyLex sLex;
-	int nByte;
-	sxi32 rc;
-	if( nArg < 1 || !ph7_value_is_string(apArg[0]) ){
-		/* Missing/Invalid arguments, return NULL */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Extract the JSON string */
-	zIn = ph7_value_to_string(apArg[0],&nByte);
-	if( nByte < 1 ){
-		/* Empty string,return NULL */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Clear JSON error code */
-	pVm->json_rc = JSON_ERROR_NONE;
-	/* Tokenize the input */
-	SySetInit(&sToken,&pVm->sAllocator,sizeof(SyToken));
-	SyLexInit(&sLex,&sToken,VmJsonTokenize,&pVm->json_rc);
-	SyLexTokenizeInput(&sLex,zIn,(sxu32)nByte,0,0,0);
-	if( pVm->json_rc != JSON_ERROR_NONE ){
-		/* Something goes wrong while tokenizing input. [i.e: Unexpected token] */
-		SyLexRelease(&sLex);
-		SySetRelease(&sToken);
-		/* return NULL */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Fill the decoder */
-	sDecoder.pCtx = pCtx;
-	sDecoder.pErr = &pVm->json_rc;
-	sDecoder.pIn = (SyToken *)SySetBasePtr(&sToken);
-	sDecoder.pEnd = &sDecoder.pIn[SySetUsed(&sToken)];
-	sDecoder.iFlags = 0;
-	if( nArg > 1 && ph7_value_to_bool(apArg[1]) != 0 ){
-		/* Returned objects will be converted into associative arrays */
-		sDecoder.iFlags |= JSON_DECODE_ASSOC;
-	}
-	sDecoder.rec_depth = 32;
-	if( nArg > 2 && ph7_value_is_int(apArg[2]) ){
-		int nDepth = ph7_value_to_int(apArg[2]);
-		if( nDepth > 1 && nDepth < 32 ){
-			sDecoder.rec_depth = nDepth;
-		}
-	}
-	sDecoder.rec_count = 0;
-	/* Set a default consumer */
-	sDecoder.xConsumer = VmJsonDefaultDecoder;
-	sDecoder.pUserData = 0;
-	/* Decode the raw JSON input */
-	rc = VmJsonDecode(&sDecoder,0);
-	if( rc == SXERR_ABORT ||  pVm->json_rc != JSON_ERROR_NONE ){
-		/*
-		 * Something goes wrong while decoding JSON input.Return NULL.
-		 */
-		ph7_result_null(pCtx);
-	}
-	/* Clean-up the mess left behind */
-	SyLexRelease(&sLex);
-	SySetRelease(&sToken);
-	/* All done */
-	return PH7_OK;
-}
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-/*
- * XML processing Functions.
- * Status:
- *    Devel.
- */
-enum ph7_xml_handler_id{
-	PH7_XML_START_TAG = 0, /* Start element handlers ID */
-	PH7_XML_END_TAG,       /* End element handler ID*/
-	PH7_XML_CDATA,         /* Character data handler ID*/
-	PH7_XML_PI,            /* Processing instruction (PI) handler ID*/
-	PH7_XML_DEF,           /* Default handler ID */
-	PH7_XML_UNPED,         /* Unparsed entity declaration handler */
-	PH7_XML_ND,            /* Notation declaration handler ID*/
-	PH7_XML_EER,           /* External entity reference handler */
-	PH7_XML_NS_START,      /* Start namespace declaration handler */
-	PH7_XML_NS_END         /* End namespace declaration handler */
-};
-#define XML_TOTAL_HANDLER (PH7_XML_NS_END + 1)
-/* An instance of the following structure describe a working
- * XML engine instance.
- */
-typedef struct ph7_xml_engine ph7_xml_engine;
-struct ph7_xml_engine
-{
-	ph7_vm *pVm;         /* VM that own this instance */
-	ph7_context *pCtx;   /* Call context */
-	SyXMLParser sParser; /* Underlying XML parser */
-	ph7_value aCB[XML_TOTAL_HANDLER]; /* User-defined callbacks */
-	ph7_value sParserValue; /* ph7_value holding this instance which is forwarded
-							  * as the first argument to the user callbacks.
-							  */
-	int ns_sep;      /* Namespace separator */
-	SyBlob sErr;     /* Error message consumer */
-	sxi32 iErrCode;  /* Last error code */
-	sxi32 iNest;     /* Nesting level */
-	sxu32 nLine;     /* Last processed line */
-	sxu32 nMagic;    /* Magic number so that we avoid misuse  */
-};
-#define XML_ENGINE_MAGIC 0x851EFC52
-#define IS_INVALID_XML_ENGINE(XML) (XML == 0 || (XML)->nMagic != XML_ENGINE_MAGIC)
-/*
- * Allocate and initialize an XML engine.
- */
-static ph7_xml_engine * VmCreateXMLEngine(ph7_context *pCtx,int process_ns,int ns_sep)
-{
-	ph7_xml_engine *pEngine;
-	ph7_vm *pVm = pCtx->pVm;
-	ph7_value *pValue;
-	sxu32 n;
-	/* Allocate a new instance */
-	pEngine = (ph7_xml_engine *)SyMemBackendAlloc(&pVm->sAllocator,sizeof(ph7_xml_engine));
-	if( pEngine == 0 ){
-		/* Out of memory */
-		return 0;
-	}
-	/* Zero the structure */
-	SyZero(pEngine,sizeof(ph7_xml_engine));
-	/* Initialize fields */
-	pEngine->pVm = pVm;
-	pEngine->pCtx = 0;
-	pEngine->ns_sep = ns_sep;
-	SyXMLParserInit(&pEngine->sParser,&pVm->sAllocator,process_ns ? SXML_ENABLE_NAMESPACE : 0);
-	SyBlobInit(&pEngine->sErr,&pVm->sAllocator);
-	PH7_MemObjInit(pVm,&pEngine->sParserValue);
-	for( n = 0 ; n < SX_ARRAYSIZE(pEngine->aCB) ; ++n ){
-		pValue = &pEngine->aCB[n];
-		/* NULLIFY the array entries,until someone register an event handler */
-		PH7_MemObjInit(&(*pVm),pValue);
-	}
-	ph7_value_resource(&pEngine->sParserValue,pEngine);
-	pEngine->iErrCode = SXML_ERROR_NONE;
-	/* Finally set the magic number */
-	pEngine->nMagic = XML_ENGINE_MAGIC;
-	return pEngine;
-}
-/*
- * Release an XML engine.
- */
-static void VmReleaseXMLEngine(ph7_xml_engine *pEngine)
-{
-	ph7_vm *pVm = pEngine->pVm;
-	ph7_value *pValue;
-	sxu32 n;
-	/* Release fields */
-	SyBlobRelease(&pEngine->sErr);
-	SyXMLParserRelease(&pEngine->sParser);
-	PH7_MemObjRelease(&pEngine->sParserValue);
-	for( n = 0 ; n < SX_ARRAYSIZE(pEngine->aCB) ; ++n ){
-		pValue = &pEngine->aCB[n];
-		PH7_MemObjRelease(pValue);
-	}
-	pEngine->nMagic = 0x2621;
-	/* Finally,release the whole instance */
-	SyMemBackendFree(&pVm->sAllocator,pEngine);
-}
-/*
- * resource xml_parser_create([ string $encoding ])
- *  Create an UTF-8 XML parser.
- * Parameter
- *  $encoding
- *   (Only UTF-8 encoding is used)
- * Return
- *  Returns a resource handle for the new XML parser.
- */
-static int vm_builtin_xml_parser_create(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	/* Allocate a new instance */
-	pEngine = VmCreateXMLEngine(&(*pCtx),0,':');
-	if( pEngine == 0 ){
-		ph7_context_throw_error(pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-		/* Return null */
-		ph7_result_null(pCtx);
-		SXUNUSED(nArg); /* cc warning */
-		SXUNUSED(apArg);
-		return PH7_OK;
-	}
-	/* Return the engine as a resource */
-	ph7_result_resource(pCtx,pEngine);
-	return PH7_OK;
-}
-/*
- * resource xml_parser_create_ns([ string $encoding[,string $separator = ':']])
- *  Create an UTF-8 XML parser with namespace support.
- * Parameter
- *  $encoding
- *   (Only UTF-8 encoding is supported)
- *  $separtor
- *   Namespace separator (a single character)
- * Return
- *  Returns a resource handle for the new XML parser.
- */
-static int vm_builtin_xml_parser_create_ns(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	int ns_sep = ':';
-	if( nArg > 1 && ph7_value_is_string(apArg[1]) ){
-		const char *zSep = ph7_value_to_string(apArg[1],0);
-		if( zSep[0] != 0 ){
-			ns_sep = zSep[0];
-		}
-	}
-	/* Allocate a new instance */
-	pEngine = VmCreateXMLEngine(&(*pCtx),TRUE,ns_sep);
-	if( pEngine == 0 ){
-		ph7_context_throw_error(pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-		/* Return null */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Return the engine as a resource */
-	ph7_result_resource(pCtx,pEngine);
-	return PH7_OK;
-}
-/*
- * bool xml_parser_free(resource $parser)
- *  Release an XML engine.
- * Parameter
- *  $parser
- *   A reference to the XML parser to free.
- * Return
- *  This function returns FALSE if parser does not refer
- *  to a valid parser, or else it frees the parser and returns TRUE.
- */
-static int vm_builtin_xml_parser_free(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Safely release the engine */
-	VmReleaseXMLEngine(pEngine);
-	/* Return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_element_handler(resource $parser,callback $start_element_handler,[callback $end_element_handler])
- * Sets the element handler functions for the XML parser. start_element_handler and end_element_handler
- * are strings containing the names of functions.
- * Parameters
- *  $parser
- *   A reference to the XML parser to set up start and end element handler functions.
- *  $start_element_handler
- *    The function named by start_element_handler must accept three parameters:
- *    start_element_handler(resource $parser,string $name,array $attribs)
- *    $parser
- *      The first parameter, parser, is a reference to the XML parser calling the handler.
- *   $name
- *      The second parameter, name, contains the name of the element for which this handler
- *		is called.If case-folding is in effect for this parser, the element name will be in uppercase letters.
- *  $attribs
- *      The third parameter, attribs, contains an associative array with the element's attributes (if any).
- *		The keys of this array are the attribute names, the values are the attribute values.
- *      Attribute names are case-folded on the same criteria as element names.Attribute values are not case-folded.
- *      The original order of the attributes can be retrieved by walking through attribs the normal way, using each().
- *      The first key in the array was the first attribute, and so on.
- *      Note: Instead of a function name, an array containing an object reference and a method name can also be supplied.
- * $end_element_handler
- *     The function named by end_element_handler must accept two parameters:
- *     end_element_handler(resource $parser,string $name)
- *    $parser
- *      The first parameter, parser, is a reference to the XML parser calling the handler.
- *   $name
- *      The second parameter, name, contains the name of the element for which this handler
- *      is called.If case-folding is in effect for this parser, the element name will be in uppercase
- *      letters.
- *      If a handler function is set to an empty string, or FALSE, the handler in question is disabled.
- * Return
- * TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_element_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the start_element_handler callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_START_TAG]);
-		if( nArg > 2 ){
-			/* Save the end_element_handler callback for later invocation */
-			PH7_MemObjStore(apArg[2]/* User callback*/,&pEngine->aCB[PH7_XML_END_TAG]);
-		}
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_character_data_handler(resource $parser,callback $handler)
- *  Sets the character data handler function for the XML parser parser.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept two parameters:
- *   handler(resource $parser,string $data)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $data
- *   The second parameter, data, contains the character data as a string.
- *   Character data handler is called for every piece of a text in the XML document.
- *   It can be called multiple times inside each fragment (e.g. for non-ASCII strings).
- *   If a handler function is set to an empty string, or FALSE, the handler in question is disabled.
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_character_data_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_CDATA]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_default_handler(resource $parser,callback $handler)
- *  Set up default handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept two parameters:
- *   handler(resource $parser,string $data)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $data
- *   The second parameter, data, contains the character data.This may be the XML declaration
- *   document type declaration, entities or other data for which no other handler exists.
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_default_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_DEF]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_end_namespace_decl_handler(resource $parser,callback $handler)
- *  Set up end namespace declaration handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept two parameters:
- *   handler(resource $parser,string $prefix)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $prefix
- *   The prefix is a string used to reference the namespace within an XML object.
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_end_namespace_decl_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_NS_END]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_start_namespace_decl_handler(resource $parser,callback $handler)
- *  Set up start namespace declaration handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept two parameters:
- *   handler(resource $parser,string $prefix,string $uri)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $prefix
- *   The prefix is a string used to reference the namespace within an XML object.
- *  $uri
- *    Uniform Resource Identifier (URI) of namespace.
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_start_namespace_decl_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_NS_START]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_processing_instruction_handler(resource $parser,callback $handler)
- *  Set up processing instruction (PI) handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept three parameters:
- *   handler(resource $parser,string $target,string $data)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $target
- *   The second parameter, target, contains the PI target.
- *  $data
-     The third parameter, data, contains the PI data.
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_processing_instruction_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_PI]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_unparsed_entity_decl_handler(resource $parser,callback $handler)
- *  Set up unparsed entity declaration handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept six parameters:
- *  handler(resource $parser,string $entity_name,string $base,string $system_id,string $public_id,string $notation_name)
- *  $parser
- *   The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $entity_name
- *   The name of the entity that is about to be defined.
- *  $base
- *   This is the base for resolving the system identifier (systemId) of the external entity.
- *   Currently this parameter will always be set to an empty string.
- *  $system_id
- *   System identifier for the external entity.
- *  $public_id
- *    Public identifier for the external entity.
- *  $notation_name
- *    Name of the notation of this entity (see xml_set_notation_decl_handler()).
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_unparsed_entity_decl_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_UNPED]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_notation_decl_handler(resource $parser,callback $handler)
- *  Set up notation declaration handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept five parameters:
- *  handler(resource $parser,string $entity_name,string $base,string $system_id,string $public_id)
- *  $parser
- *   The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $entity_name
- *   The name of the entity that is about to be defined.
- *  $base
- *   This is the base for resolving the system identifier (systemId) of the external entity.
- *   Currently this parameter will always be set to an empty string.
- *  $system_id
- *   System identifier for the external entity.
- *  $public_id
- *    Public identifier for the external entity.
- *  Note: Instead of a function name, an array containing an object reference and a method name
- *  can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_notation_decl_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_ND]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * bool xml_set_external_entity_ref_handler(resource $parser,callback $handler)
- *  Set up external entity reference handler.
- * Parameters
- * $parser
- *   A reference to the XML parser to set up character data handler function.
- * $handler
- *  handler is a string containing the name of the callback.
- *  The function named by handler must accept five parameters:
- *   handler(resource $parser,string $open_entity_names,string $base,string $system_id,string $public_id)
- *  $parser
- *   The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $open_entity_names
- *   The second parameter, open_entity_names, is a space-separated list of the names
- *   of the entities that are open for the parse of this entity (including the name of the referenced entity).
- *  $base
- *   This is the base for resolving the system identifier (system_id) of the external entity.
- *   Currently this parameter will always be set to an empty string.
- *  $system_id
- *   The fourth parameter, system_id, is the system identifier as specified in the entity declaration.
- *  $public_id
- *   The fifth parameter, public_id, is the public identifier as specified in the entity declaration
- *   or an empty string if none was specified; the whitespace in the public identifier will have been
- *   normalized as required by the XML spec.
- * Note: Instead of a function name, an array containing an object reference and a method name
- * can also be supplied.
- * Return
- *  TRUE on success or FALSE on failure.
- */
-static int vm_builtin_xml_set_external_entity_ref_handler(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( nArg > 1 ){
-		/* Save the user callback for later invocation */
-		PH7_MemObjStore(apArg[1]/* User callback*/,&pEngine->aCB[PH7_XML_EER]);
-	}
-	/* All done,return TRUE */
-	ph7_result_bool(pCtx,1);
-	return PH7_OK;
-}
-/*
- * int xml_get_current_line_number(resource $parser)
- *  Gets the current line number for the given XML parser.
- * Parameters
- * $parser
- *   A reference to the XML parser.
- * Return
- *  This function returns FALSE if parser does not refer
- *  to a valid parser, or else it returns which line the parser
- *  is currently at in its data buffer.
- */
-static int vm_builtin_xml_get_current_line_number(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Return the line number */
-	ph7_result_int(pCtx,(int)pEngine->nLine);
-	return PH7_OK;
-}
-/*
- * int xml_get_current_byte_index(resource $parser)
- *  Gets the current byte index of the given XML parser.
- * Parameters
- * $parser
- *   A reference to the XML parser.
- * Return
- *  This function returns FALSE if parser does not refer to a valid
- *  parser, or else it returns which byte index the parser is currently
- *  at in its data buffer (starting at 0).
- */
-static int vm_builtin_xml_get_current_byte_index(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	SyStream *pStream;
-	SyToken *pToken;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the current processed token */
-	pToken = (SyToken *)SySetPeekCurrentEntry(&pEngine->sParser.sToken);
-	if( pToken == 0 ){
-		/* Stream not yet processed */
-		ph7_result_int(pCtx,0);
-		return 0;
-	}
-	/* Point to the input stream */
-	pStream = &pEngine->sParser.sLex.sStream;
-	/* Return the byte index */
-	ph7_result_int64(pCtx,(ph7_int64)(pToken->sData.zString-(const char *)pStream->zInput));
-	return PH7_OK;
-}
-/*
- * bool xml_set_object(resource $parser,object &$object)
- *  Use XML Parser within an object.
- * NOTE
- *  This function is depreceated and is a no-op.
- * Parameters
- * $parser
- *   A reference to the XML parser.
- * $object
- *  The object where to use the XML parser.
- * Return
- * Always FALSE.
- */
-static int vm_builtin_xml_set_object(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 2 || !ph7_value_is_resource(apArg[0]) || !ph7_value_is_object(apArg[1]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/*  Throw a notice and return */
-	ph7_context_throw_error(pCtx,PH7_CTX_NOTICE,"This function is depreceated and is a no-op."
-		"In order to mimic this behaviour,you can supply instead of a function name an array "
-		"containing an object reference and a method name."
-		);
-	/* Return FALSE */
-	ph7_result_bool(pCtx,0);
-	return PH7_OK;
-}
-/*
- * int xml_get_current_column_number(resource $parser)
- *  Gets the current column number of the given XML parser.
- * Parameters
- * $parser
- *   A reference to the XML parser.
- * Return
- *  This function returns FALSE if parser does not refer to a valid parser, or else it returns
- *  which column on the current line (as given by xml_get_current_line_number()) the parser
- *  is currently at.
- */
-static int vm_builtin_xml_get_current_column_number(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	SyStream *pStream;
-	SyToken *pToken;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the current processed token */
-	pToken = (SyToken *)SySetPeekCurrentEntry(&pEngine->sParser.sToken);
-	if( pToken == 0 ){
-		/* Stream not yet processed */
-		ph7_result_int(pCtx,0);
-		return 0;
-	}
-	/* Point to the input stream */
-	pStream = &pEngine->sParser.sLex.sStream;
-	/* Return the byte index */
-	ph7_result_int64(pCtx,(ph7_int64)(pToken->sData.zString-(const char *)pStream->zInput)/80);
-	return PH7_OK;
-}
-/*
- * int xml_get_error_code(resource $parser)
- *  Get XML parser error code.
- * Parameters
- * $parser
- *   A reference to the XML parser.
- * Return
- *  This function returns FALSE if parser does not refer to a valid
- *  parser, or else it returns one of the error codes listed in the error
- *  codes section.
- */
-static int vm_builtin_xml_get_error_code(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 1 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Return the error code if any */
-	ph7_result_int(pCtx,pEngine->iErrCode);
-	return PH7_OK;
-}
-/*
- * XML parser event callbacks
- * Each time the unserlying XML parser extract a single token
- * from the input,one of the following callbacks are invoked.
- * IMP-XML-ENGINE-07-07-2012 22:02 FreeBSD [chm@symisc.net]
- */
-/*
- * Create a scalar ph7_value holding the value
- * of an XML tag/attribute/CDATA and so on.
- */
-static ph7_value * VmXMLValue(ph7_xml_engine *pEngine,SyXMLRawStr *pXML,SyXMLRawStr *pNsUri)
-{
-	ph7_value *pValue;
-	/* Allocate a new scalar variable */
-	pValue = ph7_context_new_scalar(pEngine->pCtx);
-	if( pValue == 0 ){
-		ph7_context_throw_error(pEngine->pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-		return 0;
-	}
-	if( pNsUri && pNsUri->nByte > 0 ){
-		/* Append namespace URI and the separator */
-		ph7_value_string_format(pValue,"%.*s%c",pNsUri->nByte,pNsUri->zString,pEngine->ns_sep);
-	}
-	/* Copy the tag value */
-	ph7_value_string(pValue,pXML->zString,(int)pXML->nByte);
-	return pValue;
-}
-/*
- * Create a 'ph7_value' of type array holding the values
- * of an XML tag attributes.
- */
-static ph7_value * VmXMLAttrValue(ph7_xml_engine *pEngine,SyXMLRawStr *aAttr,sxu32 nAttr)
-{
-	ph7_value *pArray;
-	/* Create an empty array */
-	pArray = ph7_context_new_array(pEngine->pCtx);
-	if( pArray == 0 ){
-		ph7_context_throw_error(pEngine->pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-		return 0;
-	}
-	if( nAttr > 0 ){
-		ph7_value *pKey,*pValue;
-		sxu32 n;
-		/* Create worker variables */
-		pKey = ph7_context_new_scalar(pEngine->pCtx);
-		pValue = ph7_context_new_scalar(pEngine->pCtx);
-		if( pKey == 0 || pValue == 0 ){
-			ph7_context_throw_error(pEngine->pCtx,PH7_CTX_ERR,"PH7 is running out of memory");
-			return 0;
-		}
-		/* Copy attributes */
-		for( n = 0 ; n < nAttr ; n += 2 ){
-			/* Reset string cursors */
-			ph7_value_reset_string_cursor(pKey);
-			ph7_value_reset_string_cursor(pValue);
-			/* Copy attribute name and it's associated value */
-			ph7_value_string(pKey,aAttr[n].zString,(int)aAttr[n].nByte); /* Attribute name */
-			ph7_value_string(pValue,aAttr[n+1].zString,(int)aAttr[n+1].nByte); /* Attribute value */
-			/* Insert in the array */
-			ph7_array_add_elem(pArray,pKey,pValue); /* Will make it's own copy */
-		}
-		/* Release the worker variables */
-		ph7_context_release_value(pEngine->pCtx,pKey);
-		ph7_context_release_value(pEngine->pCtx,pValue);
-	}
-	/* Return the freshly created array */
-	return pArray;
-}
-/*
- * Start element handler.
- * The user defined callback must accept three parameters:
- *    start_element_handler(resource $parser,string $name,array $attribs )
- *    $parser
- *      The first parameter, parser, is a reference to the XML parser calling the handler.
- *    $name
- *      The second parameter, name, contains the name of the element for which this handler
- *		is called.If case-folding is in effect for this parser, the element name will be in uppercase letters.
- *    $attribs
- *      The third parameter, attribs, contains an associative array with the element's attributes (if any).
- *		The keys of this array are the attribute names, the values are the attribute values.
- *      Attribute names are case-folded on the same criteria as element names.Attribute values are not case-folded.
- *      The original order of the attributes can be retrieved by walking through attribs the normal way, using each().
- *      The first key in the array was the first attribute, and so on.
- *      Note: Instead of a function name, an array containing an object reference and a method name can also be supplied.
- */
-static sxi32 VmXMLStartElementHandler(SyXMLRawStr *pStart,SyXMLRawStr *pNS,sxu32 nAttr,SyXMLRawStr *aAttr,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	ph7_value *pCallback,*pTag,*pAttr;
-	/* Point to the target user defined callback */
-	pCallback = &pEngine->aCB[PH7_XML_START_TAG];
-	/* Make sure the given callback is callable */
-	if( !PH7_VmIsCallable(pEngine->pVm,pCallback,0) ){
-		/* Not callable,return immediately*/
-		return SXRET_OK;
-	}
-	/* Create a ph7_value holding the tag name */
-	pTag = VmXMLValue(pEngine,pStart,pNS);
-	/* Create a ph7_value holding the tag attributes */
-	pAttr = VmXMLAttrValue(pEngine,aAttr,nAttr);
-	if( pTag == 0  || pAttr == 0 ){
-		SXUNUSED(pNS); /* cc warning */
-		/* Out of mem,return immediately */
-		return SXRET_OK;
-	}
-	/* Invoke the user callback */
-	PH7_VmCallUserFunctionAp(pEngine->pVm,pCallback,0,&pEngine->sParserValue,pTag,pAttr,(ph7_value*)0);
-	/* Clean-up the mess left behind */
-	ph7_context_release_value(pEngine->pCtx,pTag);
-	ph7_context_release_value(pEngine->pCtx,pAttr);
-	return SXRET_OK;
-}
-/*
- * End element handler.
- * The user defined callback must accept two parameters:
- *  end_element_handler(resource $parser,string $name)
- *  $parser
- *   The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $name
- *   The second parameter, name, contains the name of the element for which this handler is called.
- *   If case-folding is in effect for this parser, the element name will be in uppercase letters.
- *   Note: Instead of a function name, an array containing an object reference and a method name
- *   can also be supplied.
- */
-static sxi32 VmXMLEndElementHandler(SyXMLRawStr *pEnd,SyXMLRawStr *pNS,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	ph7_value *pCallback,*pTag;
-	/* Point to the target user defined callback */
-	pCallback = &pEngine->aCB[PH7_XML_END_TAG];
-	/* Make sure the given callback is callable */
-	if( !PH7_VmIsCallable(pEngine->pVm,pCallback,0) ){
-		/* Not callable,return immediately*/
-		return SXRET_OK;
-	}
-	/* Create a ph7_value holding the tag name */
-	pTag = VmXMLValue(pEngine,pEnd,pNS);
-	if( pTag == 0  ){
-		SXUNUSED(pNS); /* cc warning */
-		/* Out of mem,return immediately */
-		return SXRET_OK;
-	}
-	/* Invoke the user callback */
-	PH7_VmCallUserFunctionAp(pEngine->pVm,pCallback,0,&pEngine->sParserValue,pTag,(ph7_value*)0);
-	/* Clean-up the mess left behind */
-	ph7_context_release_value(pEngine->pCtx,pTag);
-	return SXRET_OK;
-}
-/*
- * Character data handler.
- *  The user defined callback must accept two parameters:
- *  handler(resource $parser,string $data)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $data
- *   The second parameter, data, contains the character data as a string.
- *   Character data handler is called for every piece of a text in the XML document.
- *   It can be called multiple times inside each fragment (e.g. for non-ASCII strings).
- *   If a handler function is set to an empty string, or FALSE, the handler in question is disabled.
- *   Note: Instead of a function name, an array containing an object reference and a method name can also be supplied.
- */
-static sxi32 VmXMLTextHandler(SyXMLRawStr *pText,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	ph7_value *pCallback,*pData;
-	/* Point to the target user defined callback */
-	pCallback = &pEngine->aCB[PH7_XML_CDATA];
-	/* Make sure the given callback is callable */
-	if( !PH7_VmIsCallable(pEngine->pVm,pCallback,0) ){
-		/* Not callable,return immediately*/
-		return SXRET_OK;
-	}
-	/* Create a ph7_value holding the data */
-	pData = VmXMLValue(pEngine,&(*pText),0);
-	if( pData == 0  ){
-		/* Out of mem,return immediately */
-		return SXRET_OK;
-	}
-	/* Invoke the user callback */
-	PH7_VmCallUserFunctionAp(pEngine->pVm,pCallback,0,&pEngine->sParserValue,pData,(ph7_value*)0);
-	/* Clean-up the mess left behind */
-	ph7_context_release_value(pEngine->pCtx,pData);
-	return SXRET_OK;
-}
-/*
- * Processing instruction (PI) handler.
- * The user defined callback must accept two parameters:
- *   handler(resource $parser,string $target,string $data)
- *  $parser
- *    The first parameter, parser, is a reference to the XML parser calling the handler.
- *  $target
- *   The second parameter, target, contains the PI target.
- *  $data
- *    The third parameter, data, contains the PI data.
- *    Note: Instead of a function name, an array containing an object reference
- *    and a method name can also be supplied.
- */
-static sxi32 VmXMLPIHandler(SyXMLRawStr *pTargetStr,SyXMLRawStr *pDataStr,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	ph7_value *pCallback,*pTarget,*pData;
-	/* Point to the target user defined callback */
-	pCallback = &pEngine->aCB[PH7_XML_PI];
-	/* Make sure the given callback is callable */
-	if( !PH7_VmIsCallable(pEngine->pVm,pCallback,0) ){
-		/* Not callable,return immediately*/
-		return SXRET_OK;
-	}
-	/* Get a ph7_value holding the data */
-	pTarget = VmXMLValue(pEngine,&(*pTargetStr),0);
-	pData = VmXMLValue(pEngine,&(*pDataStr),0);
-	if( pTarget == 0 || pData == 0  ){
-		/* Out of mem,return immediately */
-		return SXRET_OK;
-	}
-	/* Invoke the user callback */
-	PH7_VmCallUserFunctionAp(pEngine->pVm,pCallback,0,&pEngine->sParserValue,pTarget,pData,(ph7_value*)0);
-	/* Clean-up the mess left behind */
-	ph7_context_release_value(pEngine->pCtx,pTarget);
-	ph7_context_release_value(pEngine->pCtx,pData);
-	return SXRET_OK;
-}
-/*
- * Namespace declaration handler.
- * The user defined callback must accept two parameters:
- *    handler(resource $parser,string $prefix,string $uri)
- * $parser
- *   The first parameter, parser, is a reference to the XML parser calling the handler.
- * $prefix
- *   The prefix is a string used to reference the namespace within an XML object.
- * $uri
- *   Uniform Resource Identifier (URI) of namespace.
- *   Note: Instead of a function name, an array containing an object reference
- *   and a method name can also be supplied.
- */
-static sxi32 VmXMLNSStartHandler(SyXMLRawStr *pUriStr,SyXMLRawStr *pPrefixStr,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	ph7_value *pCallback,*pUri,*pPrefix;
-	/* Point to the target user defined callback */
-	pCallback = &pEngine->aCB[PH7_XML_NS_START];
-	/* Make sure the given callback is callable */
-	if( !PH7_VmIsCallable(pEngine->pVm,pCallback,0) ){
-		/* Not callable,return immediately*/
-		return SXRET_OK;
-	}
-	/* Get a ph7_value holding the PREFIX/URI */
-	pUri = VmXMLValue(pEngine,pUriStr,0);
-	pPrefix = VmXMLValue(pEngine,pPrefixStr,0);
-	if( pUri == 0 || pPrefix == 0  ){
-		/* Out of mem,return immediately */
-		return SXRET_OK;
-	}
-	/* Invoke the user callback */
-	PH7_VmCallUserFunctionAp(pEngine->pVm,pCallback,0,&pEngine->sParserValue,pUri,pPrefix,(ph7_value*)0);
-	/* Clean-up the mess left behind */
-	ph7_context_release_value(pEngine->pCtx,pUri);
-	ph7_context_release_value(pEngine->pCtx,pPrefix);
-	return SXRET_OK;
-}
-/*
- * Namespace end declaration handler.
- * The user defined callback must accept two parameters:
- *    handler(resource $parser,string $prefix)
- * $parser
- *   The first parameter, parser, is a reference to the XML parser calling the handler.
- * $prefix
- *  The prefix is a string used to reference the namespace within an XML object.
- *   Note: Instead of a function name, an array containing an object reference
- *   and a method name can also be supplied.
- */
-static sxi32 VmXMLNSEndHandler(SyXMLRawStr *pPrefixStr,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	ph7_value *pCallback,*pPrefix;
-	/* Point to the target user defined callback */
-	pCallback = &pEngine->aCB[PH7_XML_NS_END];
-	/* Make sure the given callback is callable */
-	if( !PH7_VmIsCallable(pEngine->pVm,pCallback,0) ){
-		/* Not callable,return immediately*/
-		return SXRET_OK;
-	}
-	/* Get a ph7_value holding the prefix */
-	pPrefix = VmXMLValue(pEngine,pPrefixStr,0);
-	if( pPrefix == 0 ){
-		/* Out of mem,return immediately */
-		return SXRET_OK;
-	}
-	/* Invoke the user callback */
-	PH7_VmCallUserFunctionAp(pEngine->pVm,pCallback,0,&pEngine->sParserValue,pPrefix,(ph7_value*)0);
-	/* Clean-up the mess left behind */
-	ph7_context_release_value(pEngine->pCtx,pPrefix);
-	return SXRET_OK;
-}
-/*
- * Error Message consumer handler.
- * Each time the XML parser encounter a syntaxt error or any other error
- * related to XML processing,the following callback is invoked by the
- * underlying XML parser.
- */
-static sxi32 VmXMLErrorHandler(const char *zMessage,sxi32 iErrCode,SyToken *pToken,void *pUserData)
-{
-	ph7_xml_engine *pEngine = (ph7_xml_engine *)pUserData;
-	/* Save the error code */
-	pEngine->iErrCode = iErrCode;
-	SXUNUSED(zMessage); /* cc warning */
-	if( pToken ){
-		pEngine->nLine = pToken->nLine;
-	}
-	/* Abort XML processing immediately */
-	return SXERR_ABORT;
-}
-/*
- * int xml_parse(resource $parser,string $data[,bool $is_final = false ])
- *  Parses an XML document. The handlers for the configured events are called
- *  as many times as necessary.
- * Parameters
- *  $parser
- *   A reference to the XML parser.
- *  $data
- *   Chunk of data to parse. A document may be parsed piece-wise by calling
- *   xml_parse() several times with new data, as long as the is_final parameter
- *   is set and TRUE when the last data is parsed.
- * $is_final
- *   NOT USED. This implementation require that all the processed input be
- *   entirely loaded in memory.
- * Return
- *  Returns 1 on success or 0 on failure.
- */
-static int vm_builtin_xml_parse(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	SyXMLParser *pParser;
-	const char *zData;
-	int nByte;
-	if( nArg < 2 || !ph7_value_is_resource(apArg[0]) || !ph7_value_is_string(apArg[1]) ){
-		/* Missing/Ivalid arguments,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	if( pEngine->iNest > 0 ){
-		/* This can happen when the user callback call xml_parse() again
-		 * in it's body which is forbidden.
-		 */
-		ph7_context_throw_error_format(pCtx,PH7_CTX_ERR,
-			"Recursive call to %s,PH7 is returning false",
-			ph7_function_name(pCtx)
-			);
-		/* Return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	pEngine->pCtx = pCtx;
-	/* Point to the underlying XML parser */
-	pParser = &pEngine->sParser;
-	/* Register elements handler */
-	SyXMLParserSetEventHandler(pParser,pEngine,
-		VmXMLStartElementHandler,
-		VmXMLTextHandler,
-		VmXMLErrorHandler,
-		0,
-		VmXMLEndElementHandler,
-		VmXMLPIHandler,
-		0,
-		0,
-		VmXMLNSStartHandler,
-		VmXMLNSEndHandler
-		);
-	pEngine->iErrCode = SXML_ERROR_NONE;
-	/* Extract the raw XML input */
-	zData = ph7_value_to_string(apArg[1],&nByte);
-	/* Start the parse process */
-	pEngine->iNest++;
-	SyXMLProcess(pParser,zData,(sxu32)nByte);
-	pEngine->iNest--;
-	/* Return the parse result */
-	ph7_result_int(pCtx,pEngine->iErrCode == SXML_ERROR_NONE ? 1 : 0);
-	return PH7_OK;
-}
-/*
- * bool xml_parser_set_option(resource $parser,int $option,mixed $value)
- *  Sets an option in an XML parser.
- * Parameters
- *  $parser
- *   A reference to the XML parser to set an option in.
- *  $option
- *    Which option to set. See below.
- *   The following options are available:
- *   XML_OPTION_CASE_FOLDING 	integer  Controls whether case-folding is enabled for this XML parser.
- *   XML_OPTION_SKIP_TAGSTART 	integer  Specify how many characters should be skipped in the beginning of a tag name.
- *   XML_OPTION_SKIP_WHITE 	    integer  Whether to skip values consisting of whitespace characters.
- *   XML_OPTION_TARGET_ENCODING string 	 Sets which target encoding to use in this XML parser.
- * $value
- *   The option's new value.
- * Return
- *  Returns 1 on success or 0 on failure.
- * Note:
- *  Well,none of these options have meaning under the built-in XML parser so a call to this
- *  function is a no-op.
- */
-static int vm_builtin_xml_parser_set_option(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	if( nArg < 2 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Always return FALSE */
-	ph7_result_bool(pCtx,0);
-	return PH7_OK;
-}
-/*
- * mixed xml_parser_get_option(resource $parser,int $option)
- *  Get options from an XML parser.
- * Parameters
- *  $parser
- *   A reference to the XML parser to set an option in.
- * $option
- *   Which option to fetch.
- * Return
- *  This function returns FALSE if parser does not refer to a valid parser
- *  or if option isn't valid.Else the option's value is returned.
- */
-static int vm_builtin_xml_parser_get_option(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	ph7_xml_engine *pEngine;
-	int nOp;
-	if( nArg < 2 || !ph7_value_is_resource(apArg[0]) ){
-		/* Missing/Ivalid argument,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Point to the XML engine */
-	pEngine = (ph7_xml_engine *)ph7_value_to_resource(apArg[0]);
-	if( IS_INVALID_XML_ENGINE(pEngine) ){
-		/* Corrupt engine,return FALSE */
-		ph7_result_bool(pCtx,0);
-		return PH7_OK;
-	}
-	/* Extract the option */
-	nOp = ph7_value_to_int(apArg[1]);
-	switch(nOp){
-	case SXML_OPTION_SKIP_TAGSTART:
-	case SXML_OPTION_SKIP_WHITE:
-	case SXML_OPTION_CASE_FOLDING:
-		ph7_result_int(pCtx,0); break;
-	case SXML_OPTION_TARGET_ENCODING:
-		ph7_result_string(pCtx,"UTF-8",(int)sizeof("UTF-8")-1);
-		break;
-	default:
-		/* Unknown option,return FALSE*/
-		ph7_result_bool(pCtx,0);
-		break;
-	}
-	return PH7_OK;
-}
-/*
- * string xml_error_string(int $code)
- *  Gets the XML parser error string associated with the given code.
- * Parameters
- *  $code
- *   An error code from xml_get_error_code().
- * Return
- *  Returns a string with a textual description of the error
- *  code, or FALSE if no description was found.
- */
-static int vm_builtin_xml_error_string(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	int nErr = -1;
-	if( nArg > 0 ){
-		nErr = ph7_value_to_int(apArg[0]);
-	}
-	switch(nErr){
-	case SXML_ERROR_DUPLICATE_ATTRIBUTE:
-		ph7_result_string(pCtx,"Duplicate attribute",-1/*Compute length automatically*/);
-		break;
-	case SXML_ERROR_INCORRECT_ENCODING:
-		ph7_result_string(pCtx,"Incorrect encoding",-1);
-		break;
-	case SXML_ERROR_INVALID_TOKEN:
-		ph7_result_string(pCtx,"Unexpected token",-1);
-		break;
-	case SXML_ERROR_MISPLACED_XML_PI:
-		ph7_result_string(pCtx,"Misplaced processing instruction",-1);
-		break;
-	case SXML_ERROR_NO_MEMORY:
-		ph7_result_string(pCtx,"Out of memory",-1);
-		break;
-	case SXML_ERROR_NONE:
-		ph7_result_string(pCtx,"Not an error",-1);
-		break;
-	case SXML_ERROR_TAG_MISMATCH:
-		ph7_result_string(pCtx,"Tag mismatch",-1);
-		break;
-	case -1:
-		ph7_result_string(pCtx,"Unknown error code",-1);
-		break;
-	default:
-		ph7_result_string(pCtx,"Syntax error",-1);
-		break;
-	}
-	return PH7_OK;
-}
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-/*
- * int utf8_encode(string $input)
- *  UTF-8 encoding.
- *  This function encodes the string data to UTF-8, and returns the encoded version.
- *  UTF-8 is a standard mechanism used by Unicode for encoding wide character values
- * into a byte stream. UTF-8 is transparent to plain ASCII characters, is self-synchronized
- * (meaning it is possible for a program to figure out where in the bytestream characters start)
- * and can be used with normal string comparison functions for sorting and such.
- *  Notes on UTF-8 (According to SQLite3 authors):
- *  Byte-0    Byte-1    Byte-2    Byte-3    Value
- *  0xxxxxxx                                 00000000 00000000 0xxxxxxx
- *  110yyyyy  10xxxxxx                       00000000 00000yyy yyxxxxxx
- *  1110zzzz  10yyyyyy  10xxxxxx             00000000 zzzzyyyy yyxxxxxx
- *  11110uuu  10uuzzzz  10yyyyyy  10xxxxxx   000uuuuu zzzzyyyy yyxxxxxx
- * Parameters
- * $input
- *   String to encode or NULL on failure.
- * Return
- *  An UTF-8 encoded string.
- */
-static int vm_builtin_utf8_encode(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	const unsigned char *zIn,*zEnd;
-	int nByte,c,e;
-	if( nArg < 1 ){
-		/* Missing arguments,return null */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Extract the target string */
-	zIn = (const unsigned char *)ph7_value_to_string(apArg[0],&nByte);
-	if( nByte < 1 ){
-		/* Empty string,return null */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	zEnd = &zIn[nByte];
-	/* Start the encoding process */
-	for(;;){
-		if( zIn >= zEnd ){
-			/* End of input */
-			break;
-		}
-		c = zIn[0];
-		/* Advance the stream cursor */
-		zIn++;
-		/* Encode */
-		if( c<0x00080 ){
-			e = (c&0xFF);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-		}else if( c<0x00800 ){
-			e = 0xC0 + ((c>>6)&0x1F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-			e = 0x80 + (c & 0x3F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-		}else if( c<0x10000 ){
-			e = 0xE0 + ((c>>12)&0x0F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-			e = 0x80 + ((c>>6) & 0x3F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-			e = 0x80 + (c & 0x3F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-		}else{
-			e = 0xF0 + ((c>>18) & 0x07);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-			e = 0x80 + ((c>>12) & 0x3F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-			e = 0x80 + ((c>>6) & 0x3F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-			e = 0x80 + (c & 0x3F);
-			ph7_result_string(pCtx,(const char *)&e,(int)sizeof(char));
-		}
-	}
-	/* All done */
-	return PH7_OK;
-}
-/*
- * UTF-8 decoding routine extracted from the sqlite3 source tree.
- * Original author: D. Richard Hipp (http://www.sqlite.org)
- * Status: Public Domain
- */
-/*
-** This lookup table is used to help decode the first byte of
-** a multi-byte UTF8 character.
-*/
-static const unsigned char UtfTrans1[] = {
-  0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
-  0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
-  0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17,
-  0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
-  0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
-  0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
-  0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
-  0x00, 0x01, 0x02, 0x03, 0x00, 0x01, 0x00, 0x00,
-};
-/*
-** Translate a single UTF-8 character.  Return the unicode value.
-**
-** During translation, assume that the byte that zTerm points
-** is a 0x00.
-**
-** Write a pointer to the next unread byte back into *pzNext.
-**
-** Notes On Invalid UTF-8:
-**
-**  *  This routine never allows a 7-bit character (0x00 through 0x7f) to
-**     be encoded as a multi-byte character.  Any multi-byte character that
-**     attempts to encode a value between 0x00 and 0x7f is rendered as 0xfffd.
-**
-**  *  This routine never allows a UTF16 surrogate value to be encoded.
-**     If a multi-byte character attempts to encode a value between
-**     0xd800 and 0xe000 then it is rendered as 0xfffd.
-**
-**  *  Bytes in the range of 0x80 through 0xbf which occur as the first
-**     byte of a character are interpreted as single-byte characters
-**     and rendered as themselves even though they are technically
-**     invalid characters.
-**
-**  *  This routine accepts an infinite number of different UTF8 encodings
-**     for unicode values 0x80 and greater.  It do not change over-length
-**     encodings to 0xfffd as some systems recommend.
-*/
-#define READ_UTF8(zIn, zTerm, c)                           \
-  c = *(zIn++);                                            \
-  if( c>=0xc0 ){                                           \
-    c = UtfTrans1[c-0xc0];                                 \
-    while( zIn!=zTerm && (*zIn & 0xc0)==0x80 ){            \
-      c = (c<<6) + (0x3f & *(zIn++));                      \
-    }                                                      \
-    if( c<0x80                                             \
-        || (c&0xFFFFF800)==0xD800                          \
-        || (c&0xFFFFFFFE)==0xFFFE ){  c = 0xFFFD; }        \
-  }
-PH7_PRIVATE int PH7_Utf8Read(
-  const unsigned char *z,         /* First byte of UTF-8 character */
-  const unsigned char *zTerm,     /* Pretend this byte is 0x00 */
-  const unsigned char **pzNext    /* Write first byte past UTF-8 char here */
-){
-  int c;
-  READ_UTF8(z, zTerm, c);
-  *pzNext = z;
-  return c;
-}
-/*
- * string utf8_decode(string $data)
- *  This function decodes data, assumed to be UTF-8 encoded, to unicode.
- * Parameters
- * data
- *  An UTF-8 encoded string.
- * Return
- *  Unicode decoded string or NULL on failure.
- */
-static int vm_builtin_utf8_decode(ph7_context *pCtx,int nArg,ph7_value **apArg)
-{
-	const unsigned char *zIn,*zEnd;
-	int nByte,c;
-	if( nArg < 1 ){
-		/* Missing arguments,return null */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	/* Extract the target string */
-	zIn = (const unsigned char *)ph7_value_to_string(apArg[0],&nByte);
-	if( nByte < 1 ){
-		/* Empty string,return null */
-		ph7_result_null(pCtx);
-		return PH7_OK;
-	}
-	zEnd = &zIn[nByte];
-	/* Start the decoding process */
-	while( zIn < zEnd ){
-		c = PH7_Utf8Read(zIn,zEnd,&zIn);
-		if( c == 0x0 ){
-			break;
-		}
-		ph7_result_string(pCtx,(const char *)&c,(int)sizeof(char));
-	}
-	return PH7_OK;
-}
+/* JSON encoding/decoding routines moved to vm_json.c */
+/* XML processing and UTF-8 routines moved to vm_xml.c */
 /* Table of built-in VM functions. */
 static const ph7_builtin_func aVmFunc[] = {
 	{ "func_num_args"  , vm_builtin_func_num_args },
@@ -14457,761 +12116,4 @@ PH7_PRIVATE const ph7_io_stream * PH7_VmGetStreamDevice(
 	return 0;
 }
 #endif /* PH7_DISABLE_BUILTIN_FUNC || PH7_DISABLE_DISK_IO */
-/*
- * Section:
- *    HTTP/URI related routines.
- * Status:
- *    Stable.
- */
- /*
-  * URI Parser: Split an URI into components [i.e: Host,Path,Query,...].
-  * URI syntax: [method:/][/[user[:pwd]@]host[:port]/][document]
-  * This almost, but not quite, RFC1738 URI syntax.
-  * This routine is not a validator,it does not check for validity
-  * nor decode URI parts,the only thing this routine does is splitting
-  * the input to its fields.
-  * Upper layer are responsible of decoding and validating URI parts.
-  * On success,this function populate the "SyhttpUri" structure passed
-  * as the first argument. Otherwise SXERR_* is returned when a malformed
-  * input is encountered.
-  */
- static sxi32 VmHttpSplitURI(SyhttpUri *pOut,const char *zUri,sxu32 nLen)
- {
-	 const char *zEnd = &zUri[nLen];
-	 sxu8 bHostOnly = FALSE;
-	 sxu8 bIPv6 = FALSE	;
-	 const char *zCur;
-	 SyString *pComp;
-	 sxu32 nPos = 0;
-	 sxi32 rc;
-	 /* Zero the structure first */
-	 SyZero(pOut,sizeof(SyhttpUri));
-	 /* Remove leading and trailing white spaces  */
-	 SyStringInitFromBuf(&pOut->sRaw,zUri,nLen);
-	 SyStringFullTrim(&pOut->sRaw);
-	 /* Find the first '/' separator */
-	 rc = SyByteFind(zUri,(sxu32)(zEnd - zUri),'/',&nPos);
-	 if( rc != SXRET_OK ){
-		 /* Assume a host name only */
-		 zCur = zEnd;
-		 bHostOnly = TRUE;
-		 goto ProcessHost;
-	 }
-	 zCur = &zUri[nPos];
-	 if( zUri != zCur && zCur[-1] == ':' ){
-		 /* Extract a scheme:
-		  * Not that we can get an invalid scheme here.
-		  * Fortunately the caller can discard any URI by comparing this scheme with its
-		  * registered schemes and will report the error as soon as his comparison function
-		  * fail.
-		  */
-	 	pComp = &pOut->sScheme;
-		SyStringInitFromBuf(pComp,zUri,(sxu32)(zCur - zUri - 1));
-		SyStringLeftTrim(pComp);
-	 }
-	 if( zCur[1] != '/' ){
-		 if( zCur == zUri || zCur[-1] == ':' ){
-		  /* No authority */
-		  goto PathSplit;
-		}
-		 /* There is something here , we will assume its an authority
-		  * and someone has forgot the two prefix slashes "//",
-		  * sooner or later we will detect if we are dealing with a malicious
-		  * user or not,but now assume we are dealing with an authority
-		  * and let the caller handle all the validation process.
-		  */
-		 goto ProcessHost;
-	 }
-	 zUri = &zCur[2];
-	 zCur = zEnd;
-	 rc = SyByteFind(zUri,(sxu32)(zEnd - zUri),'/',&nPos);
-	 if( rc == SXRET_OK ){
-		 zCur = &zUri[nPos];
-	 }
- ProcessHost:
-	 /* Extract user information if present */
-	 rc = SyByteFind(zUri,(sxu32)(zCur - zUri),'@',&nPos);
-	 if( rc == SXRET_OK ){
-		 if( nPos > 0 ){
-			 sxu32 nPassOfft; /* Password offset */
-			 pComp = &pOut->sUser;
-			 SyStringInitFromBuf(pComp,zUri,nPos);
-			 /* Extract the password if available */
-			 rc = SyByteFind(zUri,(sxu32)(zCur - zUri),':',&nPassOfft);
-			 if( rc == SXRET_OK && nPassOfft < nPos){
-				 pComp->nByte = nPassOfft;
-				 pComp = &pOut->sPass;
-				 pComp->zString = &zUri[nPassOfft+sizeof(char)];
-				 pComp->nByte = nPos - nPassOfft - 1;
-			 }
-			 /* Update the cursor */
-			 zUri = &zUri[nPos+1];
-		 }else{
-			 zUri++;
-		 }
-	 }
-	 pComp = &pOut->sHost;
-	 while( zUri < zCur && SyisSpace(zUri[0])){
-		 zUri++;
-	 }
-	 SyStringInitFromBuf(pComp,zUri,(sxu32)(zCur - zUri));
-	 if( pComp->zString[0] == '[' ){
-		 /* An IPv6 Address: Make a simple naive test
-		  */
-		 zUri++; pComp->zString++; pComp->nByte = 0;
-		 while( ((unsigned char)zUri[0] < 0xc0 && SyisHex(zUri[0])) || zUri[0] == ':' ){
-			 zUri++; pComp->nByte++;
-		 }
-		 if( zUri[0] != ']' ){
-			 return SXERR_CORRUPT; /* Malformed IPv6 address */
-		 }
-		 zUri++;
-		 bIPv6 = TRUE;
-	 }
-	 /* Extract a port number if available */
-	 rc = SyByteFind(zUri,(sxu32)(zCur - zUri),':',&nPos);
-	 if( rc == SXRET_OK ){
-		 if( bIPv6 == FALSE ){
-			 pComp->nByte = (sxu32)(&zUri[nPos] - zUri);
-		 }
-		 pComp = &pOut->sPort;
-		 SyStringInitFromBuf(pComp,&zUri[nPos+1],(sxu32)(zCur - &zUri[nPos+1]));
-	 }
-	 if( bHostOnly == TRUE ){
-		 return SXRET_OK;
-	 }
-PathSplit:
-	 zUri = zCur;
-	 pComp = &pOut->sPath;
-	 SyStringInitFromBuf(pComp,zUri,(sxu32)(zEnd-zUri));
-	 if( pComp->nByte == 0 ){
-		 return SXRET_OK; /* Empty path */
-	 }
-	 if( SXRET_OK == SyByteFind(zUri,(sxu32)(zEnd-zUri),'?',&nPos) ){
-		 pComp->nByte = nPos; /* Update path length */
-		 pComp = &pOut->sQuery;
-		 SyStringInitFromBuf(pComp,&zUri[nPos+1],(sxu32)(zEnd-&zUri[nPos+1]));
-	 }
-	 if( SXRET_OK == SyByteFind(zUri,(sxu32)(zEnd-zUri),'#',&nPos) ){
-		 /* Update path or query length */
-		 if( pComp == &pOut->sPath ){
-			 pComp->nByte = nPos;
-		 }else{
-			 if( &zUri[nPos] < (char *)SyStringData(pComp) ){
-				 /* Malformed syntax : Query must be present before fragment */
-				 return SXERR_SYNTAX;
-			 }
-			 pComp->nByte -= (sxu32)(zEnd - &zUri[nPos]);
-		 }
-		 pComp = &pOut->sFragment;
-		 SyStringInitFromBuf(pComp,&zUri[nPos+1],(sxu32)(zEnd-&zUri[nPos+1]))
-	 }
-	 return SXRET_OK;
- }
- /*
- * Extract a single line from a raw HTTP request.
- * Return SXRET_OK on success,SXERR_EOF when end of input
- * and SXERR_MORE when more input is needed.
- */
-static sxi32 VmGetNextLine(SyString *pCursor,SyString *pCurrent)
-{
-  	const char *zIn;
-  	sxu32 nPos;
-	/* Jump leading white spaces */
-	SyStringLeftTrim(pCursor);
-	if( pCursor->nByte < 1 ){
-		SyStringInitFromBuf(pCurrent,0,0);
-		return SXERR_EOF; /* End of input */
-	}
-	zIn = SyStringData(pCursor);
-	if( SXRET_OK != SyByteListFind(pCursor->zString,pCursor->nByte,"\r\n",&nPos) ){
-		/* Line not found,tell the caller to read more input from source */
-		SyStringDupPtr(pCurrent,pCursor);
-		return SXERR_MORE;
-	}
-  	pCurrent->zString = zIn;
-  	pCurrent->nByte	= nPos;
-  	/* advance the cursor so we can call this routine again */
-  	pCursor->zString = &zIn[nPos];
-  	pCursor->nByte -= nPos;
-  	return SXRET_OK;
- }
- /*
-  * Split a single MIME header into a name value pair.
-  * This function return SXRET_OK,SXERR_CONTINUE on success.
-  * Otherwise SXERR_NEXT is returned when a malformed header
-  * is encountered.
-  * Note: This function handle also mult-line headers.
-  */
- static sxi32 VmHttpProcessOneHeader(SyhttpHeader *pHdr,SyhttpHeader *pLast,const char *zLine,sxu32 nLen)
- {
-	 SyString *pName;
-	 sxu32 nPos;
-	 sxi32 rc;
-	 if( nLen < 1 ){
-		 return SXERR_NEXT;
-	 }
-	 /* Check for multi-line header */
-	if( pLast && (zLine[-1] == ' ' || zLine[-1] == '\t') ){
-		SyString *pTmp = &pLast->sValue;
-		SyStringFullTrim(pTmp);
-		if( pTmp->nByte == 0 ){
-			SyStringInitFromBuf(pTmp,zLine,nLen);
-		}else{
-			/* Update header value length */
-			pTmp->nByte = (sxu32)(&zLine[nLen] - pTmp->zString);
-		}
-		 /* Simply tell the caller to reset its states and get another line */
-		 return SXERR_CONTINUE;
-	 }
-	/* Split the header */
-	pName = &pHdr->sName;
-	rc = SyByteFind(zLine,nLen,':',&nPos);
-	if(rc != SXRET_OK ){
-		return SXERR_NEXT; /* Malformed header;Check the next entry */
-	}
-	SyStringInitFromBuf(pName,zLine,nPos);
-	SyStringFullTrim(pName);
-	/* Extract a header value */
-	SyStringInitFromBuf(&pHdr->sValue,&zLine[nPos + 1],nLen - nPos - 1);
-	/* Remove leading and trailing whitespaces */
-	SyStringFullTrim(&pHdr->sValue);
-	return SXRET_OK;
- }
- /*
-  * Extract all MIME headers associated with a HTTP request.
-  * After processing the first line of a HTTP request,the following
-  * routine is called in order to extract MIME headers.
-  * This function return SXRET_OK on success,SXERR_MORE when it needs
-  * more inputs.
-  * Note: Any malformed header is simply discarded.
-  */
- static sxi32 VmHttpExtractHeaders(SyString *pRequest,SySet *pOut)
- {
-	 SyhttpHeader *pLast = 0;
-	 SyString sCurrent;
-	 SyhttpHeader sHdr;
-	 sxu8 bEol;
-	 sxi32 rc;
-	 if( SySetUsed(pOut) > 0 ){
-		 pLast = (SyhttpHeader *)SySetAt(pOut,SySetUsed(pOut)-1);
-	 }
-	 bEol = FALSE;
-	 for(;;){
-		 SyZero(&sHdr,sizeof(SyhttpHeader));
-		 /* Extract a single line from the raw HTTP request */
-		 rc = VmGetNextLine(pRequest,&sCurrent);
-		 if(rc != SXRET_OK ){
-			 if( sCurrent.nByte < 1 ){
-				 break;
-			 }
-			 bEol = TRUE;
-		 }
-		 /* Process the header */
-		 if( SXRET_OK == VmHttpProcessOneHeader(&sHdr,pLast,sCurrent.zString,sCurrent.nByte)){
-			 if( SXRET_OK != SySetPut(pOut,(const void *)&sHdr) ){
-				 break;
-			 }
-			 /* Retrieve the last parsed header so we can handle multi-line header
-			  * in case we face one of them.
-			  */
-			 pLast = (SyhttpHeader *)SySetPeek(pOut);
-		 }
-		 if( bEol ){
-			 break;
-		 }
-	 } /* for(;;) */
-	 return SXRET_OK;
- }
- /*
-  * Process the first line of a HTTP request.
-  * This routine perform the following operations
-  *  1) Extract the HTTP method.
-  *  2) Split the request URI to it's fields [ie: host,path,query,...].
-  *  3) Extract the HTTP protocol version.
-  */
- static sxi32 VmHttpProcessFirstLine(
-	 SyString *pRequest, /* Raw HTTP request */
-	 sxi32 *pMethod,     /* OUT: HTTP method */
-	 SyhttpUri *pUri,    /* OUT: Parse of the URI */
-	 sxi32 *pProto       /* OUT: HTTP protocol */
-	 )
- {
-	 static const char *azMethods[] = { "get","post","head","put"};
-	 static const sxi32 aMethods[]  = { HTTP_METHOD_GET,HTTP_METHOD_POST,HTTP_METHOD_HEAD,HTTP_METHOD_PUT};
-	 const char *zIn,*zEnd,*zPtr;
-	 SyString sLine;
-	 sxu32 nLen;
-	 sxi32 rc;
-	 /* Extract the first line and update the pointer */
-	 rc = VmGetNextLine(pRequest,&sLine);
-	 if( rc != SXRET_OK ){
-		 return rc;
-	 }
-	 if ( sLine.nByte < 1 ){
-		 /* Empty HTTP request */
-		 return SXERR_EMPTY;
-	 }
-	 /* Delimit the line and ignore trailing and leading white spaces */
-	 zIn = sLine.zString;
-	 zEnd = &zIn[sLine.nByte];
-	 while( zIn < zEnd && (unsigned char)zIn[0] < 0xc0 && SyisSpace(zIn[0]) ){
-		 zIn++;
-	 }
-	 /* Extract the HTTP method */
-	 zPtr = zIn;
-	 while( zIn < zEnd && !SyisSpace(zIn[0]) ){
-		 zIn++;
-	 }
-	 *pMethod = HTTP_METHOD_OTHR;
-	 if( zIn > zPtr ){
-		 sxu32 i;
-		 nLen = (sxu32)(zIn-zPtr);
-		 for( i = 0 ; i < SX_ARRAYSIZE(azMethods) ; ++i ){
-			 if( SyStrnicmp(azMethods[i],zPtr,nLen) == 0 ){
-				 *pMethod = aMethods[i];
-				 break;
-			 }
-		 }
-	 }
-	 /* Jump trailing white spaces */
-	 while( zIn < zEnd && (unsigned char)zIn[0] < 0xc0 && SyisSpace(zIn[0]) ){
-		 zIn++;
-	 }
-	  /* Extract the request URI */
-	 zPtr = zIn;
-	 while( zIn < zEnd && !SyisSpace(zIn[0]) ){
-		 zIn++;
-	 }
-	 if( zIn > zPtr ){
-		 nLen = (sxu32)(zIn-zPtr);
-		 /* Split raw URI to it's fields */
-		 VmHttpSplitURI(pUri,zPtr,nLen);
-	 }
-	 /* Jump trailing white spaces */
-	 while( zIn < zEnd && (unsigned char)zIn[0] < 0xc0 && SyisSpace(zIn[0]) ){
-		 zIn++;
-	 }
-	 /* Extract the HTTP version */
-	 zPtr = zIn;
-	 while( zIn < zEnd && !SyisSpace(zIn[0]) ){
-		 zIn++;
-	 }
-	 *pProto = HTTP_PROTO_11; /* HTTP/1.1 */
-	 rc = 1;
-	 if( zIn > zPtr ){
-		 rc = SyStrnicmp(zPtr,"http/1.0",(sxu32)(zIn-zPtr));
-	 }
-	 if( !rc ){
-		 *pProto = HTTP_PROTO_10; /* HTTP/1.0 */
-	 }
-	 return SXRET_OK;
- }
- /*
-  * Tokenize,decode and split a raw query encoded as: "x-www-form-urlencoded"
-  * into a name value pair.
-  * Note that this encoding is implicit in GET based requests.
-  * After the tokenization process,register the decoded queries
-  * in the $_GET/$_POST/$_REQUEST superglobals arrays.
-  */
- static sxi32 VmHttpSplitEncodedQuery(
-	 ph7_vm *pVm,       /* Target VM */
-	 SyString *pQuery,  /* Raw query to decode */
-	 SyBlob *pWorker,   /* Working buffer */
-	 int is_post        /* TRUE if we are dealing with a POST request */
-	 )
- {
-	 const char *zEnd = &pQuery->zString[pQuery->nByte];
-	 const char *zIn = pQuery->zString;
-	 ph7_value *pGet,*pRequest;
-	 SyString sName,sValue;
-	 const char *zPtr;
-	 sxu32 nBlobOfft;
-	 /* Extract superglobals */
-	 if( is_post ){
-		 /* $_POST superglobal */
-		 pGet = VmExtractSuper(&(*pVm),"_POST",sizeof("_POST")-1);
-	 }else{
-		 /* $_GET superglobal */
-		 pGet = VmExtractSuper(&(*pVm),"_GET",sizeof("_GET")-1);
-	 }
-	 pRequest = VmExtractSuper(&(*pVm),"_REQUEST",sizeof("_REQUEST")-1);
-	 /* Split up the raw query */
-	 for(;;){
-		 /* Jump leading white spaces */
-		 while(zIn < zEnd  && SyisSpace(zIn[0]) ){
-			 zIn++;
-		 }
-		 if( zIn >= zEnd ){
-			 break;
-		 }
-		 zPtr = zIn;
-		 while( zPtr < zEnd && zPtr[0] != '=' && zPtr[0] != '&' && zPtr[0] != ';' ){
-			 zPtr++;
-		 }
-		 /* Reset the working buffer */
-		 SyBlobReset(pWorker);
-		 /* Decode the entry */
-		 SyUriDecode(zIn,(sxu32)(zPtr-zIn),PH7_VmBlobConsumer,pWorker,TRUE);
-		 /* Save the entry */
-		 sName.nByte = SyBlobLength(pWorker);
-		 sValue.zString = 0;
-		 sValue.nByte = 0;
-		 if( zPtr < zEnd && zPtr[0] == '=' ){
-			 zPtr++;
-			 zIn = zPtr;
-			 /* Store field value */
-			 while( zPtr < zEnd && zPtr[0] != '&' && zPtr[0] != ';' ){
-				 zPtr++;
-			 }
-			 if( zPtr > zIn ){
-				 /* Decode the value */
-				  nBlobOfft = SyBlobLength(pWorker);
-				  SyUriDecode(zIn,(sxu32)(zPtr-zIn),PH7_VmBlobConsumer,pWorker,TRUE);
-				  sValue.zString = (const char *)SyBlobDataAt(pWorker,nBlobOfft);
-				  sValue.nByte = SyBlobLength(pWorker) - nBlobOfft;
-
-			 }
-			 /* Synchronize pointers */
-			 zIn = zPtr;
-		 }
-		 sName.zString = (const char *)SyBlobData(pWorker);
-		 /* Install the decoded query in the $_GET/$_REQUEST array */
-		 if( pGet && (pGet->iFlags & MEMOBJ_HASHMAP) ){
-			 VmHashmapInsert((ph7_hashmap *)pGet->x.pOther,
-				 sName.zString,(int)sName.nByte,
-				 sValue.zString,(int)sValue.nByte
-				 );
-		 }
-		 if( pRequest && (pRequest->iFlags & MEMOBJ_HASHMAP) ){
-			 VmHashmapInsert((ph7_hashmap *)pRequest->x.pOther,
-				 sName.zString,(int)sName.nByte,
-				 sValue.zString,(int)sValue.nByte
-					 );
-		 }
-		 /* Advance the pointer */
-		 zIn = &zPtr[1];
-	 }
-	/* All done*/
-	return SXRET_OK;
- }
- /*
-  * Extract MIME header value from the given set.
-  * Return header value on success. NULL otherwise.
-  */
- static SyString * VmHttpExtractHeaderValue(SySet *pSet,const char *zMime,sxu32 nByte)
- {
-	 SyhttpHeader *aMime,*pMime;
-	 SyString sMime;
-	 sxu32 n;
-	 SyStringInitFromBuf(&sMime,zMime,nByte);
-	 /* Point to the MIME entries */
-	 aMime = (SyhttpHeader *)SySetBasePtr(pSet);
-	 /* Perform the lookup */
-	 for( n = 0 ; n < SySetUsed(pSet) ; ++n ){
-		 pMime = &aMime[n];
-		 if( SyStringCmp(&sMime,&pMime->sName,SyStrnicmp) == 0 ){
-			 /* Header found,return it's associated value */
-			 return &pMime->sValue;
-		 }
-	 }
-	 /* No such MIME header */
-	 return 0;
- }
- /*
-  * Tokenize and decode a raw "Cookie:" MIME header into a name value pair
-  * and insert it's fields [i.e name,value] in the $_COOKIE superglobal.
-  */
- static sxi32 VmHttpPorcessCookie(ph7_vm *pVm,SyBlob *pWorker,const char *zIn,sxu32 nByte)
- {
-	 const char *zPtr,*zDelimiter,*zEnd = &zIn[nByte];
-	 SyString sName,sValue;
-	 ph7_value *pCookie;
-	 sxu32 nOfft;
-	 /* Make sure the $_COOKIE superglobal is available */
-	 pCookie = VmExtractSuper(&(*pVm),"_COOKIE",sizeof("_COOKIE")-1);
-	 if( pCookie == 0 || (pCookie->iFlags & MEMOBJ_HASHMAP) == 0 ){
-		 /* $_COOKIE superglobal not available */
-		 return SXERR_NOTFOUND;
-	 }
-	 for(;;){
-		  /* Jump leading white spaces */
-		 while( zIn < zEnd && SyisSpace(zIn[0]) ){
-			 zIn++;
-		 }
-		 if( zIn >= zEnd ){
-			 break;
-		 }
-		  /* Reset the working buffer */
-		 SyBlobReset(pWorker);
-		 zDelimiter = zIn;
-		 /* Delimit the name[=value]; pair */
-		 while( zDelimiter < zEnd && zDelimiter[0] != ';' ){
-			 zDelimiter++;
-		 }
-		 zPtr = zIn;
-		 while( zPtr < zDelimiter && zPtr[0] != '=' ){
-			 zPtr++;
-		 }
-		 /* Decode the cookie */
-		 SyUriDecode(zIn,(sxu32)(zPtr-zIn),PH7_VmBlobConsumer,pWorker,TRUE);
-		 sName.nByte = SyBlobLength(pWorker);
-		 zPtr++;
-		 sValue.zString = 0;
-		 sValue.nByte = 0;
-		 if( zPtr < zDelimiter ){
-			 /* Got a Cookie value */
-			 nOfft = SyBlobLength(pWorker);
-			 SyUriDecode(zPtr,(sxu32)(zDelimiter-zPtr),PH7_VmBlobConsumer,pWorker,TRUE);
-			 SyStringInitFromBuf(&sValue,SyBlobDataAt(pWorker,nOfft),SyBlobLength(pWorker)-nOfft);
-		 }
-		 /* Synchronize pointers */
-		 zIn = &zDelimiter[1];
-		 /* Perform the insertion */
-		 sName.zString = (const char *)SyBlobData(pWorker);
-		 VmHashmapInsert((ph7_hashmap *)pCookie->x.pOther,
-			 sName.zString,(int)sName.nByte,
-			 sValue.zString,(int)sValue.nByte
-			 );
-	 }
-	 return SXRET_OK;
- }
- /*
-  * Process a full HTTP request and populate the appropriate arrays
-  * such as $_SERVER,$_GET,$_POST,$_COOKIE,$_REQUEST,... with the information
-  * extracted from the raw HTTP request. As an extension Symisc introduced
-  * the $_HEADER array which hold a copy of the processed HTTP MIME headers
-  * and their associated values. [i.e: $_HEADER['Server'],$_HEADER['User-Agent'],...].
-  * This function return SXRET_OK on success. Any other return value indicates
-  * a malformed HTTP request.
-  */
- static sxi32 VmHttpProcessRequest(ph7_vm *pVm,const char *zRequest,int nByte)
- {
-	 SyString *pName,*pValue,sRequest; /* Raw HTTP request */
-	 ph7_value *pHeaderArray;          /* $_HEADER superglobal (Symisc eXtension to the PHP specification)*/
-	 SyhttpHeader *pHeader;            /* MIME header */
-	 SyhttpUri sUri;     /* Parse of the raw URI*/
-	 SyBlob sWorker;     /* General purpose working buffer */
-	 SySet sHeader;      /* MIME headers set */
-	 sxi32 iMethod;      /* HTTP method [i.e: GET,POST,HEAD...]*/
-	 sxi32 iVer;         /* HTTP protocol version */
-	 sxi32 rc;
-	 SyStringInitFromBuf(&sRequest,zRequest,nByte);
-	 SySetInit(&sHeader,&pVm->sAllocator,sizeof(SyhttpHeader));
-	 SyBlobInit(&sWorker,&pVm->sAllocator);
-	 /* Ignore leading and trailing white spaces*/
-	 SyStringFullTrim(&sRequest);
-	 /* Process the first line */
-	 rc = VmHttpProcessFirstLine(&sRequest,&iMethod,&sUri,&iVer);
-	 if( rc != SXRET_OK ){
-		 return rc;
-	 }
-	 /* Process MIME headers */
-	 VmHttpExtractHeaders(&sRequest,&sHeader);
-	 /*
-	  * Setup $_SERVER environments
-	  */
-	 /* 'SERVER_PROTOCOL': Name and revision of the information protocol via which the page was requested */
-	 ph7_vm_config(pVm,
-		 PH7_VM_CONFIG_SERVER_ATTR,
-		 "SERVER_PROTOCOL",
-		 iVer == HTTP_PROTO_10 ? "HTTP/1.0" : "HTTP/1.1",
-		 sizeof("HTTP/1.1")-1
-		 );
-	 /* 'REQUEST_METHOD':  Which request method was used to access the page */
-	 ph7_vm_config(pVm,
-		 PH7_VM_CONFIG_SERVER_ATTR,
-		 "REQUEST_METHOD",
-		 iMethod == HTTP_METHOD_GET ?   "GET" :
-		 (iMethod == HTTP_METHOD_POST ? "POST":
-		 (iMethod == HTTP_METHOD_PUT  ? "PUT" :
-		 (iMethod == HTTP_METHOD_HEAD ?  "HEAD" : "OTHER"))),
-		 -1 /* Compute attribute length automatically */
-		 );
-	 if( SyStringLength(&sUri.sQuery) > 0 && iMethod == HTTP_METHOD_GET ){
-		 pValue = &sUri.sQuery;
-		 /* 'QUERY_STRING': The query string, if any, via which the page was accessed */
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "QUERY_STRING",
-			 pValue->zString,
-			 pValue->nByte
-			 );
-		 /* Decoded the raw query */
-		 VmHttpSplitEncodedQuery(&(*pVm),pValue,&sWorker,FALSE);
-	 }
-	 /* REQUEST_URI: The URI which was given in order to access this page; for instance, '/index.html' */
-	 pValue = &sUri.sRaw;
-	 ph7_vm_config(pVm,
-		 PH7_VM_CONFIG_SERVER_ATTR,
-		 "REQUEST_URI",
-		 pValue->zString,
-		 pValue->nByte
-		 );
-	 /*
-	  * 'PATH_INFO'
-	  * 'ORIG_PATH_INFO'
-      * Contains any client-provided pathname information trailing the actual script filename but preceding
-	  * the query string, if available. For instance, if the current script was accessed via the URL
-	  * http://www.example.com/php/path_info.php/some/stuff?foo=bar, then $_SERVER['PATH_INFO'] would contain
-	  * /some/stuff.
-	  */
-	 pValue = &sUri.sPath;
-	 ph7_vm_config(pVm,
-		 PH7_VM_CONFIG_SERVER_ATTR,
-		 "PATH_INFO",
-		 pValue->zString,
-		 pValue->nByte
-		 );
-	 ph7_vm_config(pVm,
-		 PH7_VM_CONFIG_SERVER_ATTR,
-		 "ORIG_PATH_INFO",
-		 pValue->zString,
-		 pValue->nByte
-		 );
-	 /* 'HTTP_ACCEPT': Contents of the Accept: header from the current request, if there is one */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Accept",sizeof("Accept")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_ACCEPT",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* 'HTTP_ACCEPT_CHARSET': Contents of the Accept-Charset: header from the current request, if there is one. */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Accept-Charset",sizeof("Accept-Charset")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_ACCEPT_CHARSET",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* 'HTTP_ACCEPT_ENCODING': Contents of the Accept-Encoding: header from the current request, if there is one. */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Accept-Encoding",sizeof("Accept-Encoding")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_ACCEPT_ENCODING",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	  /* 'HTTP_ACCEPT_LANGUAGE': Contents of the Accept-Language: header from the current request, if there is one */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Accept-Language",sizeof("Accept-Language")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_ACCEPT_LANGUAGE",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* 'HTTP_CONNECTION': Contents of the Connection: header from the current request, if there is one. */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Connection",sizeof("Connection")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_CONNECTION",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* 'HTTP_HOST': Contents of the Host: header from the current request, if there is one. */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Host",sizeof("Host")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_HOST",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* 'HTTP_REFERER': Contents of the Referer: header from the current request, if there is one. */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Referer",sizeof("Referer")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_REFERER",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* 'HTTP_USER_AGENT': Contents of the Referer: header from the current request, if there is one. */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"User-Agent",sizeof("User-Agent")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "HTTP_USER_AGENT",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	  /* 'PHP_AUTH_DIGEST': When doing Digest HTTP authentication this variable is set to the 'Authorization'
-	   * header sent by the client (which you should then use to make the appropriate validation).
-	   */
-	 pValue = VmHttpExtractHeaderValue(&sHeader,"Authorization",sizeof("Authorization")-1);
-	 if( pValue ){
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "PHP_AUTH_DIGEST",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-		 ph7_vm_config(pVm,
-			 PH7_VM_CONFIG_SERVER_ATTR,
-			 "PHP_AUTH",
-			 pValue->zString,
-			 pValue->nByte
-		 );
-	 }
-	 /* Install all clients HTTP headers in the $_HEADER superglobal */
-	 pHeaderArray = VmExtractSuper(&(*pVm),"_HEADER",sizeof("_HEADER")-1);
-	 /* Iterate throw the available MIME headers*/
-	 SySetResetCursor(&sHeader);
-	 pHeader = 0; /* stupid cc warning */
-	 while( SXRET_OK == SySetGetNextEntry(&sHeader,(void **)&pHeader) ){
-		 pName  = &pHeader->sName;
-		 pValue = &pHeader->sValue;
-		 if( pHeaderArray && (pHeaderArray->iFlags & MEMOBJ_HASHMAP)){
-			 /* Insert the MIME header and it's associated value */
-			 VmHashmapInsert((ph7_hashmap *)pHeaderArray->x.pOther,
-				 pName->zString,(int)pName->nByte,
-				 pValue->zString,(int)pValue->nByte
-				 );
-		 }
-		 if( pName->nByte == sizeof("Cookie")-1 && SyStrnicmp(pName->zString,"Cookie",sizeof("Cookie")-1) == 0
-			 && pValue->nByte > 0){
-				 /* Process the name=value pair and insert them in the $_COOKIE superglobal array */
-				 VmHttpPorcessCookie(&(*pVm),&sWorker,pValue->zString,pValue->nByte);
-		 }
-	 }
-	 if( iMethod == HTTP_METHOD_POST ){
-		 /* Extract raw POST data */
-		 pValue = VmHttpExtractHeaderValue(&sHeader,"Content-Type",sizeof("Content-Type") - 1);
-		 if( pValue && pValue->nByte >= sizeof("application/x-www-form-urlencoded") - 1 &&
-			 SyMemcmp("application/x-www-form-urlencoded",pValue->zString,pValue->nByte) == 0 ){
-				 /* Extract POST data length */
-				 pValue = VmHttpExtractHeaderValue(&sHeader,"Content-Length",sizeof("Content-Length") - 1);
-				 if( pValue ){
-					 sxi32 iLen = 0; /* POST data length */
-					 SyStrToInt32(pValue->zString,pValue->nByte,(void *)&iLen,0);
-					 if( iLen > 0 ){
-						 /* Remove leading and trailing white spaces */
-						 SyStringFullTrim(&sRequest);
-						 if( (int)sRequest.nByte > iLen ){
-							 sRequest.nByte = (sxu32)iLen;
-						 }
-						 /* Decode POST data now */
-						 VmHttpSplitEncodedQuery(&(*pVm),&sRequest,&sWorker,TRUE);
-					 }
-				 }
-		 }
-	 }
-	 /* All done,clean-up the mess left behind */
-	 SySetRelease(&sHeader);
-	 SyBlobRelease(&sWorker);
-	 return SXRET_OK;
- }
+/* HTTP/URI routines moved to vm_http.c */
